% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{add_marker_table_links}
\alias{add_marker_table_links}
\title{Convert specified columns of a marker table to HTML links.}
\usage{
add_marker_table_links(
  df,
  links = list(list(colname_target = "ENSEMBL", colname_source = "ENSEMBL", template_url
    = "https://www.ensembl.org/{ensembl_species}/Gene/Summary?g={id}"),
    list(colname_target = "ENTREZID", colname_source = "ENTREZID", template_url =
    "https://www.ncbi.nlm.nih.gov/gene/{id}"), list(colname_target = "SYMBOL",
    colname_source = "SYMBOL", template_url =
    "https://www.genecards.org/cgi-bin/carddisp.pl?gene={id}")),
  ensembl_species = "Homo_sapiens"
)
}
\arguments{
\item{df}{A dataframe-like object.}

\item{links}{A list of named lists. Each named list has three items:
\itemize{
\item \code{colname_target}: which column in \code{df} will be converted to link.
\item \code{colname_source}: which column in \code{df} will be used as data source.
\item \code{template_url}: URL template for link where \code{{id}} will be replaced by values in \code{colname_source} column of \code{df}.
For \code{ENSEMBL} column is also used the \code{ensembl_species} parameter (see the default value of \code{links}).
}}

\item{ensembl_species}{A character scalar: ID/name of ENSEMBL species.}
}
\value{
A modified \code{df}.
}
\description{
Convert specified columns of a marker table to HTML links.
}
\concept{sc_markers}
