% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce.R
\name{additional_cell_data_fn}
\alias{additional_cell_data_fn}
\title{Load additional cell data from a CSV or Rds (dataframe) file.}
\usage{
additional_cell_data_fn(additional_cell_data_file)
}
\arguments{
\item{additional_cell_data_file}{A character scalar: path to input file.}
}
\value{
A dataframe-like object.
}
\description{
Load additional cell data from a CSV or Rds (dataframe) file.
}
\concept{sc_sce}
