% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce.R
\name{as_seurat}
\alias{as_seurat}
\title{Convert a \code{SingleCellExperiment} to \code{Seurat} object.}
\usage{
as_seurat(sce, sce_assay = NULL, seurat_assay = "RNA", add_rowData = TRUE, ...)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object.}

\item{sce_assay}{A character scalar: name of assay in \code{sce} (e.g. \code{counts} or \code{logcounts}).
Use \code{NULL} to convert all assays (default).}

\item{seurat_assay}{A character scalar: name of assay in the new \code{Seurat} object.}

\item{add_rowData}{A logical scalar: if \code{TRUE}, add \code{rowData(sce)} to \code{meta.features} slot of assay of the new
\code{Seurat} object.}

\item{...}{Passed to \code{\link[Seurat:reexports]{Seurat::as.Seurat()}}.}
}
\value{
A \code{Seurat} object.
}
\description{
A wrapper around \code{\link[Seurat:reexports]{Seurat::as.Seurat()}}.
}
\concept{sc_sce}
