% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_that_}
\alias{assert_that_}
\title{A wrapper around \code{\link[assertthat:assert_that]{assertthat::assert_that()}}.}
\usage{
assert_that_(..., env = parent.frame(), msg = NULL)
}
\arguments{
\item{...}{Passed to \code{\link[assertthat:assert_that]{assertthat::assert_that()}}.}

\item{env}{Passed to \code{\link[assertthat:assert_that]{assertthat::assert_that()}} and \code{\link[cli:cli_alert]{cli::cli_alert_danger()}}. Defaults to caller env.}

\item{msg}{Passed to \code{\link[cli:cli_alert]{cli::cli_alert_danger()}}.}
}
\value{
\code{TRUE} if assertion is \code{TRUE} :)
}
\description{
Provides a nicer error message generated through \code{\link[cli:cli_alert]{cli::cli_alert_danger()}}.
}
\concept{misc_utils}
