% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{calc_sc3}
\alias{calc_sc3}
\title{Calculate SC3 clustering.}
\usage{
calc_sc3(
  sce_pca,
  sc3_k,
  sc3_dry = FALSE,
  integration = FALSE,
  BPPARAM = BiocParallel::SerialParam()
)
}
\arguments{
\item{sce_pca}{A \code{SingleCellExperiment} object with calculated PCA and known number of selected PCs.}

\item{sc3_k}{A numeric vector: numbers of clusters for SC3.}

\item{sc3_dry}{A logical scalar: if \code{TRUE}, SC3 won't be run, but random cell-cluster assignments will be made.
Used for testing as SC3 is very time-consuming.}

\item{integration}{A logical scalar: if \code{TRUE}, \code{logcounts()} will be coerced to matrix.}

\item{BPPARAM}{A \link[BiocParallel:BiocParallelParam-class]{BiocParallel::BiocParallelParam} object.}
}
\value{
\itemize{
\item If \code{sc3_dry = FALSE}: a \code{SingleCellExperiment} object as returned from \code{\link[SC3:sc3]{SC3::sc3()}}.
\item If \code{sc3_dry = TRUE}: a \code{SingleCellExperiment} object with random cell-cluster assignments mimicking output from
\code{\link[SC3:sc3]{SC3::sc3()}}.
}

The \code{sc3_dry} item is added to \code{metadata()} of the returned \code{SingleCellExperiment} object.
}
\description{
See \code{\link[SC3:sc3]{SC3::sc3()}} for details.
}
\concept{sc_clustering}
