% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_sample_norm_clustering.R
\name{cc_genes_fn}
\alias{cc_genes_fn}
\title{Get dataframe of cell cycle genes.}
\usage{
cc_genes_fn(sce_final_input_qc, organism, annotation_db_file)
}
\arguments{
\item{sce_final_input_qc}{(\emph{input target}) A \code{SingleCellExperiment} object.}

\item{organism}{A character scalar: if \code{"mouse"}, gene symbols will be converted to sentence-case (e.g. "Gapdh").}

\item{annotation_db_file}{A character scalar: path to \code{AnnotationDbi} SQLite file.}
}
\value{
A dataframe. \emph{Output target}: \code{cc_genes}
}
\description{
This is a wrapper around data bundled in the \code{Seurat} package (\code{cc.genes.updated.2019}).
}
\concept{single_sample_norm_clustering_fn}
