% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_annotation.R
\name{cell_annotation_diagnostic_plots_files_fn}
\alias{cell_annotation_diagnostic_plots_files_fn}
\title{Save cell annotation diagnostic plots to PDF files.}
\usage{
cell_annotation_diagnostic_plots_files_fn(cell_annotation_diagnostic_plots)
}
\arguments{
\item{cell_annotation_diagnostic_plots}{(\emph{input target}) A \code{tibble}.}
}
\value{
A character vector. \emph{Output target}: \code{cell_annotation_diagnostic_plots_files}
}
\description{
Save cell annotation diagnostic plots to PDF files.
}
\concept{sc_cell_annotation}
