% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_annotation.R
\name{cell_annotation_diagnostic_plots_fn}
\alias{cell_annotation_diagnostic_plots_fn}
\title{Generate diagnostic plots for cell annotation.}
\usage{
cell_annotation_diagnostic_plots_fn(
  cell_annotation,
  cell_data,
  sce,
  base_out_dir,
  cluster_cols_regex = "^cluster_"
)
}
\arguments{
\item{cell_annotation}{(\emph{input target}) A \code{tibble}.}

\item{cell_data}{(\emph{input target}) A \code{DataFrame}.}

\item{sce}{A \code{SingleCellExperiment} object.}

\item{base_out_dir}{A character scalar: path to output directory under which will be for each reference dataset its
diagnostic plots saved in.}

\item{cluster_cols_regex}{A character scalar: regex to match columns in \code{colData()} with cluster labels.
Those columns will be used in annotation score heatmaps.}
}
\value{
A \code{tibble}. \emph{Output target}: \code{cell_annotation_diagnostic_plots}
}
\description{
Three list and character vector columns are added to \code{cell_annotation}:
\itemize{
\item \code{score_heatmaps}: heatmaps of per-cell label scores created for each clustering (as column annotation)
(\href{http://bioconductor.org/books/release/SingleRBook/annotation-diagnostics.html#based-on-the-scores-within-cells}{details}).
\item \code{marker_heatmaps}: \code{NULL} if \code{cell_annotation$train_params$de} is not \code{"de"}, otherwise heatmap for each label
containing top upregulated markers from pairwise t-tests
(\href{http://bioconductor.org/books/release/SingleRBook/annotation-diagnostics.html#based-on-marker-gene-expression}{details}).
Number of top markers is specified in \code{cell_annotation$diagnostics_params$heatmap_n_top_markers}.
\item \code{delta_distribution_plot}: violin plots (in one figure / object) of per-cell deltas for each label.
Deltas are differences between the score for the assigned label and the median across all labels for each cell
(\href{http://bioconductor.org/books/release/SingleRBook/annotation-diagnostics.html#based-on-the-deltas-across-cells}{details}).
}

Columns with output PDF files are named as \code{score_heatmaps_out_file} etc.
}
\concept{sc_cell_annotation}
