% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_annotation.R
\name{cell_annotation_labels_fn}
\alias{cell_annotation_labels_fn}
\title{Create a named list of cell labels returned from \code{SingleR::SingleR()}.}
\usage{
cell_annotation_labels_fn(cell_annotation)
}
\arguments{
\item{cell_annotation}{(\emph{input target}) A \code{tibble}.}
}
\value{
A named list. \emph{Output target}: \code{cell_annotation_labels}
}
\description{
\code{SingleR::SingleR()} returns three types of cell labels. In the returned list from this function,
names of those are prefixed by a reference dataset's name and modified such that:
\itemize{
\item \code{first.labels} ("raw" labels) -> \verb{<reference_name>_labels_raw}
\item \code{labels} (fine-tuned, unpruned labels) -> \verb{<reference_name>_labels}
\item \code{pruned.labels} (fine-tuned and pruned labels) -> \verb{<reference_name>_labels_pruned}
}

Pruned labels contain \code{NA}s for low quality labels.
}
\concept{sc_cell_annotation}
