% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_annotation.R
\name{cell_annotation_params_fn}
\alias{cell_annotation_params_fn}
\title{Load a list of cell annotation references into a \code{tibble}.}
\usage{
cell_annotation_params_fn(cell_annotation_sources, biomart_dataset = NULL)
}
\arguments{
\item{cell_annotation_sources}{A list of cell annotation reference sources.}

\item{biomart_dataset}{Currently not used.}
}
\value{
A \code{tibble}. \emph{Output target}: \code{cell_annotation_params}
}
\description{
Cell annotation references are defined in \verb{02_norm_clustering.yaml} (single-sample) and
\verb{02_int_clustering.yaml} (integration) config files in \code{CELL_ANNOTATION_SOURCES} parameter.
}
\concept{sc_cell_annotation}
