% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce.R
\name{make_cell_groupings}
\alias{make_cell_groupings}
\alias{cell_data_fn}
\title{Create new cell groups based on existing ones.}
\usage{
make_cell_groupings(df, cell_groupings, do_cbind = FALSE)

cell_data_fn(col_data, clusters_all, cell_groupings)
}
\arguments{
\item{df}{A dataframe.}

\item{cell_groupings}{A named list of named lists. Names of outer list will be used for new columns in \code{df}.
The nested lists must have the following values:
\itemize{
\item \code{source_column}: a name of column in \code{df} to use.
\item \code{assignments}: a named list with values in the form \code{old_level = new_level}.
See the \code{CELL_GROUPINGS} parameter in \verb{02_norm_clustering.yaml} or \verb{02_int_clustering.yaml} config.
}}

\item{do_cbind}{A logical scalar: if \code{TRUE}, bind column-wise \code{df} and dataframe with new cell groupings.}

\item{col_data}{A dataframe.}

\item{clusters_all}{A named list.}
}
\value{
A dataframe.
}
\description{
This is basically a recoding of levels. A new column (cell group) will be based on levels in the
existing column.
}
\concept{sc_sce}
