% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce.R
\name{cell_data_fn}
\alias{cell_data_fn}
\title{Merge all cell-related data to a single DataFrame.}
\usage{
cell_data_fn(
  col_data,
  clusters_all,
  cell_annotation_labels,
  cell_groupings,
  spot_deconvolution_labels,
  manual_annotation_labels,
  additional_cell_data,
  pipeline_type = c("single_sample", "integration")
)
}
\arguments{
\item{col_data}{A dataframe.}

\item{clusters_all}{A named list.}

\item{cell_annotation_labels}{A named list.}

\item{cell_groupings}{A dataframe: see \code{\link[=make_cell_groupings]{make_cell_groupings()}}.}

\item{additional_cell_data}{A data.frame-like object with additional cell data to be joined.}

\item{pipeline_type}{A character scalar:
\itemize{
\item If \code{"single_sample"}, then \code{additional_cell_data} must contain \code{Barcode} column.
\item If \code{"integration"}, then \code{additional_cell_data} must contain \code{Barcode} and \code{batch} columns.
}}
}
\value{
A DataFrame object.
}
\description{
Merge all cell-related data to a single DataFrame.
}
\concept{sc_sce}
