% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_sample_input_qc.R
\name{cell_qc_fn}
\alias{cell_qc_fn}
\title{Return a SingleCellExperiment object with artifact spatial information}
\usage{
cell_qc_fn(sce_valid_cells, mito, ribo)
}
\arguments{
\item{sce_valid_cells}{(\emph{input target}) A \code{SingleCellExperiment} object.}

\item{mito_genes}{A index vector with position of mitochondrial genes}

\item{ribo_genes}{A index vector with position of ribosomal genes}
}
\value{
A \code{SingleCellExperiment} object.
}
\description{
Return a SingleCellExperiment object with artifact spatial information
}
\concept{single_sample_input_qc_fn}
