% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_drake_functions.R
\name{cellProximityBarplot_fn}
\alias{cellProximityBarplot_fn}
\title{Proximity Barplot}
\usage{
cellProximityBarplot_fn(
  gobject = gobject,
  CPscore = CPscore,
  out_dir = out_dir
)
}
\arguments{
\item{gobject}{A \code{Giotto} object}

\item{CPscore}{A precalculated score of closenest of selected clusters}

\item{out_dir}{An output directory to place resulted barplot of closenest}
}
\description{
A proximity barplot is a plot that describes closenest of selected clusters (inheritated information)
}
\concept{single_sample_norm_spatial}
