% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_drake_functions.R
\name{cellProximityHeatmap_fn}
\alias{cellProximityHeatmap_fn}
\title{Proximity Heatmap}
\usage{
cellProximityHeatmap_fn(
  gobject = gobject,
  CPscore = CPscore,
  out_dir = out_dir
)
}
\arguments{
\item{gobject}{A \code{Giotto} object}

\item{CPscore}{A precalculated score of closenest of selected clusters}

\item{out_dir}{An output directory to place resulted heatmap of closenest}
}
\description{
A proximity heatmap is a plot that describe closenest of selected clusters using on color code
}
\concept{single_sample_norm_spatial}
