% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R, R/sce.R
\name{cells_per_cluster_table}
\alias{cells_per_cluster_table}
\title{Get a frequency table of cell-cluster assignments.}
\usage{
cells_per_cluster_table(clusters, var_name = "Cluster")

cells_per_cluster_table(clusters, var_name = "Cluster")
}
\arguments{
\item{clusters}{A vector.}

\item{var_name}{A character scalar: name of column in the returned dataframe.}
}
\value{
A dataframe with frequency of cell-cluster assignments.

A dataframe.
}
\description{
\code{\link[janitor:tabyl]{janitor::tabyl()}} is used underhood.
}
\examples{
cells_per_cluster_table(rep(1:3, each = 2))

cells_per_cluster_table(rep(c(1, 2), each = 5))

}
\concept{sc_clustering}
\concept{sc_sce}
