% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{cells_per_cluster_table_collapsed_html}
\alias{cells_per_cluster_table_collapsed_html}
\title{Print a HTML of table collapsible by button.}
\usage{
cells_per_cluster_table_collapsed_html(
  df,
  id = get_random_strings(1, 10),
  label = "Show cells per cluster table"
)
}
\arguments{
\item{df}{A dataframe-like object.}

\item{id}{A character scalar: id attribute of the table. Used to match the button and collapsible content (the table).}

\item{label}{A character scalar: label of the button.}
}
\value{
Invisibly \code{NULL}.
}
\description{
The table is generated by \code{\link[=render_bootstrap_table]{render_bootstrap_table()}}.
}
\concept{misc_html}
