% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_pandoc}
\alias{check_pandoc}
\title{Check for \code{pandoc}'s binary.}
\usage{
check_pandoc(cache = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{cache, ...}{Passed to \code{rmarkdown::find_pandoc()}.}

\item{verbose}{A logical scalar: if \code{TRUE}, be verbose.}
}
\value{
Invisibly \code{TRUE} if the directory with \code{pandoc}'s binary exists, \code{FALSE} otherwise.
}
\description{
See the \code{RSTUDIO_PANDOC} parameter in the \code{pipeline.yaml} config (\code{vignette("config_pipeline")}).
}
\concept{checks}
