% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_pkg_installed}
\alias{check_pkg_installed}
\alias{check_qs_installed}
\alias{check_future_installed}
\alias{check_clustermq_installed}
\alias{check_future.callr_installed}
\title{Check if a package is installed and display an informative message.}
\usage{
check_pkg_installed(pkg, msg = "")

check_qs_installed()

check_future_installed()

check_clustermq_installed()

check_future.callr_installed()
}
\arguments{
\item{pkg}{A character scalar: name of package.}

\item{msg}{A character scalar: additional message to be displayed.}
}
\value{
\code{TRUE} if package is installed, \code{FALSE} otherwise.
}
\description{
Check if a package is installed and display an informative message.
}
\examples{
check_pkg_installed("utils")
check_pkg_installed("zzz")
check_qs_installed()

}
\concept{misc_utils}
