% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_scdrake}
\alias{check_scdrake}
\title{Check for selected \code{scdrake} dependencies.}
\usage{
check_scdrake(verbose = TRUE)
}
\arguments{
\item{verbose}{A logical scalar: if \code{TRUE}, be verbose.}
}
\value{
Invisibly \code{TRUE} if all checks return \code{TRUE}, \code{FALSE} otherwise.
}
\description{
This function will check for:
\itemize{
\item Packages that might not be installed by default:
\href{https://github.com/traversc/qs}{qs},
\href{https://github.com/mschubert/clustermq}{clustermq},
\href{https://future.futureverse.org}{future},
\href{https://future.callr.futureverse.org}{future.callr} (if the system doesn't support forking/multicore),
\href{https://github.com/hemberg-lab/SC3}{SC3} (if it is not installed from <github.com/gorgitko/SC3>)
\item If \href{https://pandoc.org/}{pandoc} is available.
\item If \href{https://github.com/mikefarah/yq}{yq} is available.
}
}
\concept{checks}
