% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_cli}
\alias{install_cli}
\alias{check_cli}
\title{Install or check the command line interface scripts.}
\usage{
install_cli(
  dir = NULL,
  type = c("user", "system"),
  ask = TRUE,
  verbose = TRUE,
  .dry = FALSE
)

check_cli(verbose = TRUE)
}
\arguments{
\item{dir}{A character scalar: path to directory where scripts will be copied to.
If \code{NULL}, the path will be determined based on \code{type}.}

\item{type}{A character scalar:
\itemize{
\item For \code{"user"}: install into the user's home directory under \code{.local/bin}
\item For \code{"system"}: install into \verb{/usr/local/bin}
}}

\item{ask}{A logical scalar: if \code{TRUE}, ask before copying.}

\item{verbose}{A logical scalar: if \code{TRUE}, be verbose.}

\item{.dry}{A logical scalar: if \code{TRUE}, don't copy the files and just return output paths.}
}
\value{
\code{install_cli()}: invisibly a character vector of length two: paths to installed files.

\code{check_cli()}: invisibly \code{TRUE} when the checks are successful, \code{FALSE} otherwise.
}
\description{
The scripts (\code{scdrake} shell script and \code{scdrake.R}) are bundled with the \code{scdrake} package and their paths can
be retrieved with:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{system.file("scdrake", package = "scdrake", mustWork = TRUE)
system.file("scdrake.R", package = "scdrake", mustWork = TRUE)
}\if{html}{\out{</div>}}

\code{check_cli()} is checking the presence of the \code{scdrake} CLI script in the \code{PATH} environment variable, and
whether the command \verb{$ scdrake -h} finishes successfully.
}
\concept{cli}
