% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_kmeans_kc_fn}
\alias{cluster_kmeans_kc_fn}
\title{Create a named list of k-means clusters for a selected number of clusters.}
\usage{
cluster_kmeans_kc_fn(sce_pca, kmeans_k, integration = FALSE)
}
\arguments{
\item{sce_pca}{A \code{SingleCellExperiment} object with calculated PCA and known number of selected PCs.}

\item{kmeans_k}{A numeric vector: numbers of clusters for k-means.}

\item{integration}{A logical scalar: if \code{TRUE}, a different names in the returned list will be used.}
}
\value{
A named list. The names will be:
\itemize{
\item For single-sample analysis (\code{integration = FALSE}): \verb{cluster_kmeans_kc_<k>}
\item For integration analysis (\code{integration = TRUE}): \verb{cluster_int_kmeans_kc_<k>}
}

where \verb{<k>} depends on the \code{kmeans_k} function parameter.
}
\description{
"kc" = "k custom" :)
}
\concept{sc_clustering}
