% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_markers.R
\name{cluster_markers_fn}
\alias{cluster_markers_fn}
\title{Add LFC summaries for Wilcox tests, obtained from t-test results.}
\usage{
cluster_markers_fn(cluster_markers_raw)
}
\arguments{
\item{cluster_markers_raw}{(\emph{input target}) A \code{tibble}.}
}
\value{
A \code{tibble}. \emph{Output target}: \code{cluster_markers}
}
\description{
Add LFC summaries for Wilcox tests, obtained from t-test results.
}
\concept{sc_cluster_markers}
