% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_markers.R
\name{cluster_markers_heatmaps_df_fn}
\alias{cluster_markers_heatmaps_df_fn}
\title{Create a final \code{tibble} holding parameters for cluster markers heatmaps generation.}
\usage{
cluster_markers_heatmaps_df_fn(
  cluster_markers_processed,
  cluster_markers_heatmap_params,
  out_dir
)
}
\arguments{
\item{cluster_markers_processed}{(\emph{input target}) A \code{tibble}.}

\item{cluster_markers_heatmap_params}{(\emph{input target}) A \code{tibble}.}

\item{out_dir}{A character scalar: path to output directory for heatmap PDFs.}
}
\value{
A \code{tibble}. \emph{Output target}: \code{cluster_markers_heatmaps_df}
}
\description{
Compared to the \code{cluster_markers_heatmap_params} target, the output \code{tibble} is expanded to each marker test type etc.
}
\concept{sc_cluster_markers}
