% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_markers.R
\name{cluster_markers_out_fn}
\alias{cluster_markers_out_fn}
\title{Create a final \code{dataframe} of cluster markers.}
\usage{
cluster_markers_out_fn(cluster_markers_processed)
}
\arguments{
\item{cluster_markers_processed}{(\emph{input target}) A \code{tibble}.}
}
\value{
A \code{tibble}. \emph{Output target}: \code{cluster_markers_out}
}
\description{
Create a final \code{dataframe} of cluster markers.
}
\concept{sc_cluster_markers}
