% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_markers.R
\name{cluster_markers_test_params_fn}
\alias{cluster_markers_test_params_fn}
\alias{cluster_markers_plot_params_fn}
\alias{cluster_markers_heatmap_params_fn}
\title{Extract columns with certain parameters from cluster markers sources \code{tibble}.}
\usage{
cluster_markers_test_params_fn(cluster_markers_params)

cluster_markers_plot_params_fn(cluster_markers_params)

cluster_markers_heatmap_params_fn(cluster_markers_params)
}
\arguments{
\item{cluster_markers_params}{(\emph{input target}) A \code{tibble}.}
}
\value{
A \code{tibble}. \emph{Output target}: \code{cluster_markers_test_params}

A \code{tibble}. \emph{Output target}: \code{cluster_markers_plot_params}

A \code{tibble}. \emph{Output target}: \code{cluster_markers_heatmap_params}
}
\description{
Statistical test-related parameters.

Plotting-related parameters.

Heatmap-related parameters.
}
\concept{sc_cluster_markers}
