% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_markers.R
\name{cluster_markers_params_fn}
\alias{cluster_markers_params_fn}
\title{Load a list of cluster markers sources into a \code{tibble}.}
\usage{
cluster_markers_params_fn(cluster_markers_sources, cell_data)
}
\arguments{
\item{cluster_markers_sources}{A list of cluster markers sources.}

\item{cell_data}{A list of \code{colData()} columns.}
}
\value{
A \code{tibble}. \emph{Output target}: \code{cluster_markers_params}
}
\description{
Cluster markers sources are defined in \code{cluster_markers.yaml} config file in \code{CLUSTER_MARKERS_SOURCES} parameter.
}
\concept{sc_cluster_markers}
