% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_markers.R
\name{cluster_markers_processed_fn}
\alias{cluster_markers_processed_fn}
\title{Add additional summary columns of cluster markers for comparisons with other group levels.}
\usage{
cluster_markers_processed_fn(cluster_markers)
}
\arguments{
\item{cluster_markers}{(\emph{input target}) A \code{tibble}.}
}
\value{
A \code{tibble}. \strong{Output target}: \code{cluster_markers_processed}
}
\description{
Add additional summary columns of cluster markers for comparisons with other group levels.
}
\concept{sc_cluster_markers}
