% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_sc3_cluster_stability_plots_file_fn}
\alias{cluster_sc3_cluster_stability_plots_file_fn}
\title{Save SC3 cluster stability plots to a single PDF.}
\usage{
cluster_sc3_cluster_stability_plots_file_fn(
  cluster_sc3_df,
  out_dir,
  is_integration
)
}
\arguments{
\item{cluster_sc3_df}{(\emph{input target}) A \code{tibble} returned from \code{\link[=cluster_sc3_df_fn]{cluster_sc3_df_fn()}}.}

\item{out_dir}{A character scalar: output directory.}

\item{is_integration}{A logical scalar: if \code{TRUE}, clusterings will be named as \verb{cluster_int_*}, otherwise \verb{cluster_*}.}
}
\value{
A character scalar: path to output PDF file. It will be \verb{\{out_dir\}/cluster_sc3_stability_plots_k\{k\}.pdf} where \code{k}
are \code{k}s joined by \code{-}.
}
\description{
Save SC3 cluster stability plots to a single PDF.
}
\concept{sc_clustering}
