% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_sc3_df_fn}
\alias{cluster_sc3_df_fn}
\title{Create a dataframe of SC3 clusters for a selected number of clusters.}
\usage{
cluster_sc3_df_fn(cluster_sce_sc3, sc3_k, is_integration)
}
\arguments{
\item{cluster_sce_sc3}{(\emph{input target}) A \code{SingleCellExperiment} object with calculated SC3 clustering.}

\item{sc3_k}{A numeric vector: numbers of clusters for SC3.}

\item{is_integration}{A logical scalar: if \code{TRUE}, clusterings will be named as \verb{cluster_int_*}, otherwise \verb{cluster_*}.}
}
\value{
A \code{tibble} whose columns are mostly self-explanatory, except the \code{data} column, which is of \code{list} type and
contains an another \code{tibble} with \code{cluster_stability_plot} column holding a \code{ggplot} object returned from
\code{\link[SC3:sc3_plot_cluster_stability]{SC3::sc3_plot_cluster_stability()}}.
}
\description{
This function just extracts cell-cluster membership from the \code{cluster_sce_sc3} object and generates
a cluster stability plots.
}
\concept{sc_clustering}
