% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_sc3_fn}
\alias{cluster_sc3_fn}
\title{Create a named list of SC3 clusters for a selected number of clusters.}
\usage{
cluster_sc3_fn(sce_sc3, sc3_k, integration = FALSE)
}
\arguments{
\item{sce_sc3}{A \code{SingleCellExperiment} object with calculated SC3 clustering.}

\item{sc3_k}{A numeric vector: numbers of clusters for SC3.}

\item{integration}{A logical scalar: if \code{TRUE}, a different names in the returned list will be used.}
}
\value{
A named list. The names will be:
\itemize{
\item For single-sample analysis (\code{integration = FALSE}): \verb{cluster_sc3_<k>}
\item For integration analysis (\code{integration = TRUE}): \verb{cluster_int_sc3_<k>}
}

where \verb{<k>} depends on the \code{sc3_k} function parameter.
}
\description{
Create a named list of SC3 clusters for a selected number of clusters.
}
\concept{sc_clustering}
