% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_n_clusters}
\alias{get_n_clusters}
\title{Get a number of unique clusters.}
\usage{
get_n_clusters(clustering)
}
\arguments{
\item{clustering}{An integer vector.}
}
\value{
A number of unique clusters.
}
\description{
Get a number of unique clusters.
}
\concept{sc_clustering}
