% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{collapse_ensembl_multivals}
\alias{collapse_ensembl_multivals}
\title{Collapse a character vector by \code{","}.}
\usage{
collapse_ensembl_multivals(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character scalar.
}
\description{
Used as \code{multiVals} parameter in \code{\link[AnnotationDbi:AnnotationDb-class]{AnnotationDbi::mapIds()}}.
}
\concept{sc_annotation}
