% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{contrasts_fn}
\alias{contrasts_fn}
\title{Extract contrast statistics from cluster markers tests.}
\usage{
contrasts_fn(contrasts_raw, contrasts_params)
}
\arguments{
\item{contrasts_raw}{(\emph{input target}) A \code{tibble}.}

\item{contrasts_params}{(\emph{input target}) A \code{tibble}.}
}
\value{
A \code{tibble}. \emph{Output target}: \code{contrasts}
}
\description{
Extract contrast statistics from cluster markers tests.
}
\concept{sc_contrasts}
