% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{contrasts_heatmaps_df_fn}
\alias{contrasts_heatmaps_df_fn}
\title{Create a \code{tibble} holding contrast results and heatmap parameters.}
\usage{
contrasts_heatmaps_df_fn(contrasts, contrasts_heatmap_params, out_dir)
}
\arguments{
\item{contrasts}{(\emph{input target}) A \code{tibble}.}

\item{contrasts_heatmap_params}{(\emph{input target}) A \code{tibble}.}

\item{out_dir}{A character scalar: path to base directory for saving heatmaps.}
}
\value{
A \code{tibble}. \emph{Output target}: \code{contrasts_heatmaps_df}
}
\description{
Create a \code{tibble} holding contrast results and heatmap parameters.
}
\concept{sc_contrasts}
