% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{contrasts_out_fn}
\alias{contrasts_out_fn}
\title{Format a \code{tibble} with contrast results.}
\usage{
contrasts_out_fn(contrasts)
}
\arguments{
\item{contrasts}{(\emph{input target}) A \code{tibble}.}
}
\value{
A \code{tibble}. \emph{Output target}: \code{contrasts_out}
}
\description{
Format a \code{tibble} with contrast results.
}
\concept{sc_contrasts}
