% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{contrasts_params_fn}
\alias{contrasts_params_fn}
\alias{contrasts_heatmap_params_fn}
\alias{contrasts_plot_params_fn}
\title{Prepare a \code{tibble} with parameters for contrasts tests.}
\usage{
contrasts_params_fn(contrasts_sources, cell_data)

contrasts_heatmap_params_fn(contrasts_params, contrasts_test_params)

contrasts_plot_params_fn(contrasts_params)
}
\arguments{
\item{contrasts_sources}{A named list.}

\item{cell_data}{A named list: items are columns from \code{colData()}. This is used to assign cells to groups or blocks.}

\item{contrasts_params}{(\emph{input target}) A \code{tibble}.}

\item{contrasts_test_params}{(\emph{input target}) A \code{tibble}.}
}
\value{
A \code{tibble}. \emph{Output target}: \code{contrasts_params}

A \code{tibble}. \emph{Output target}: \code{contrasts_heatmap_params}

A \code{tibble}. \emph{Output target}: \code{contrasts_plot_params}
}
\description{
Extract heatmap-related parameters.

Extract plotting-related parameters.
}
\concept{sc_contrasts}
