% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce.R
\name{convertToSCE}
\alias{convertToSCE}
\title{Rewrite \code{SpatialExperiment} object to a \code{SingleCellExperiment} object.}
\usage{
convertToSCE(spe_object)
}
\arguments{
\item{spe_object}{A \code{SpatialExperiment} object.}
}
\value{
A \code{SingleCellExperiment} object
}
\description{
Rewrite \code{SpatialExperiment} object to a \code{SingleCellExperiment} object.
}
\concept{spatial_sce}
