% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{create_a_link}
\alias{create_a_link}
\title{Generate a HTML link (\verb{<a></a>}).}
\usage{
create_a_link(
  href,
  text,
  href_rel_start = NULL,
  target = "_blank",
  do_cat = FALSE,
  ...
)
}
\arguments{
\item{href}{A character scalar: URL.}

\item{text}{A character scalar: text of the link.}

\item{href_rel_start}{A character scalar: relative start of \code{href}. See the section \emph{Relative links}.}

\item{target}{A character scalar: \code{target} parameter of \verb{<a>} tag.}

\item{do_cat}{A logical scalar: if \code{TRUE}, print (using \code{cat()}) the result before returning.}

\item{...}{Named parameters to put in the resulting \verb{<a>} tag.}
}
\value{
A character scalar.
}
\description{
Generate a HTML link (\verb{<a></a>}).
}
\section{Relative links}{

Sometimes URL in the supplied \code{href} parameter is absolute, or not relative to the output in which will be displayed.
For such cases use the \code{href_rel_start} parameter, which allows to make \code{href} relative to something.

For example, you are saving a plot to a file relative to working directory, \code{output/plots/plot.pdf}, and you want
to include link to this file in your RMarkdown output, \code{output/report.html}.
From the point of view of this HTML file, \code{output/plots/plot.pdf} obviously does not exist.
However, you can use \code{href_rel_start = "output"} to specify that the \code{href} is relative to this directory,
resulting in a link to \code{plots/plot.pdf}.
Thanks to \code{\link[fs:path_math]{fs::path_rel()}}, which is used underhood, even non-children \code{href} can be used, e.g.
if HTML is located in \code{output/reports/report.html}, than \code{href_rel_start = "output/reports"} will result in a link
to \code{../plots/plot.pdf}.
}

\examples{
create_a_link("google.com", "Google")
# If you want to reference a file relative to HTML saved in "output/report.html"
create_a_link("output/plots/plot.pdf", "Link to plot", href_rel_start = "output")
}
\concept{misc_html}
