% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{create_single_sample_dirs}
\alias{create_single_sample_dirs}
\alias{create_integration_dirs}
\title{Create a basic directory structure based on paths in config.}
\usage{
create_single_sample_dirs(cfg)

create_integration_dirs(cfg)
}
\arguments{
\item{cfg}{A \code{scdrake_list}. Obtained from \code{\link[=load_single_sample_configs]{load_single_sample_configs()}} or \code{\link[=load_integration_configs]{load_integration_configs()}}.}
}
\value{
Invisibly \code{NULL}.
}
\description{
Create a basic directory structure based on paths in config.
}
\concept{create_dirs}
