% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_dummy_plot}
\alias{create_dummy_plot}
\title{Create a blank \code{ggplot} with label.}
\usage{
create_dummy_plot(label)
}
\arguments{
\item{label}{A character scalar: text to display in the plot.}
}
\value{
A \code{ggplot2} object.
}
\description{
Create a blank \code{ggplot} with label.
}
\concept{misc_utils}
