% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_cell_annotation.R
\name{create_signature_matrix_fn}
\alias{create_signature_matrix_fn}
\title{Create signature matrix from provided file containing names with markers.}
\usage{
create_signature_matrix_fn(markers_file)
}
\arguments{
\item{markers_file}{A csv file containing list of annotation names with selected markers.}
}
\description{
Create signature matrix from provided file containing names with markers.
}
\concept{manual_annotation}
