% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_drake_functions.R
\name{createGiotto_fn}
\alias{createGiotto_fn}
\title{Create Giotto object}
\usage{
createGiotto_fn(sce, annotation = FALSE, selected_clustering = NULL)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object}

\item{annotation}{A logical scalar, wheter to add additional information}

\item{selected_clustering}{Name of selected clustering/if any, to add as additional information}
}
\value{
Return a \code{Giotto} object with data precalculated using a \code{SingleCellExperiment}
}
\description{
This is a function to create Giotto experiment object from \code{SingleCellExperiment} with precalculated data
}
\concept{single_sample_norm_spatial}
