% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{dimred_plot_other_vars_params_fn}
\alias{dimred_plot_other_vars_params_fn}
\title{Make a tibble with parameters for dimred plots of selected variables.}
\usage{
dimred_plot_other_vars_params_fn(dimred_names, dimred_plots_other)
}
\arguments{
\item{dimred_names}{A character vector: names of dimreds.}

\item{dimred_plots_other}{A named list: see \code{NORM_CLUSTERING_REPORT_DIMRED_PLOTS_OTHER} parameter in
\verb{02_norm_clustering.yaml} or \verb{02_int_clustering.yaml} config.}
}
\value{
A tibble. \emph{Output target}: \code{dimred_plots_other_vars_params}
}
\description{
Make a tibble with parameters for dimred plots of selected variables.
}
\concept{sce_visualization}
