% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{dimred_plots_cell_annotation_params_df_fn}
\alias{dimred_plots_cell_annotation_params_df_fn}
\title{Make a tibble with parameters for dimred plots of cell annotation labels.}
\usage{
dimred_plots_cell_annotation_params_df_fn(
  dimred_names,
  cell_annotation_params,
  out_dir = NULL
)
}
\arguments{
\item{dimred_names}{A character vector: names of dimreds.}

\item{cell_annotation_params}{(\emph{input target}) A tibble.}

\item{out_dir}{A character scalar: output directory in which PDF and PNG files will be later saved.}
}
\value{
A tibble. \emph{Output target}: \code{dimred_plots_other_vars_params}
}
\description{
Make a tibble with parameters for dimred plots of cell annotation labels.
}
\concept{sce_visualization}
