% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{dimred_plots_clustering_united_files_fn}
\alias{dimred_plots_clustering_united_files_fn}
\title{Put clustering dimred plots for different parameters (resolution, \code{k}) into a single PDF.}
\usage{
dimred_plots_clustering_united_files_fn(
  dimred_plots_clustering,
  out_dir = NULL
)
}
\arguments{
\item{dimred_plots_clustering}{(\emph{input target}) A tibble.}

\item{out_dir}{A character scalar: output directory in which PDF files will be later saved.}
}
\value{
A tibble. \emph{Output target}: \code{dimred_plots_clustering_united_files}
}
\description{
Put clustering dimred plots for different parameters (resolution, \code{k}) into a single PDF.
}
\concept{sce_visualization}
