% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{dimred_plots_from_params_df}
\alias{dimred_plots_from_params_df}
\title{Make dimred plots of selected variables.}
\usage{
dimred_plots_from_params_df(sce_dimred, dimred_plots_params_df)
}
\arguments{
\item{sce_dimred}{(\emph{input target}) A \code{SingleCellExperiment} object. Input target in the integration plan is
\code{sce_int_clustering_final}.}

\item{dimred_plots_params_df}{(\emph{input target}) A tibble.}
}
\value{
A tibble. \emph{Output target}: \code{dimred_plots_other_vars}
}
\description{
Make dimred plots of selected variables.
}
\concept{sce_visualization}
