% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{dimred_plots_other_vars_fn}
\alias{dimred_plots_other_vars_fn}
\title{Make a dimred plot of selected variables.}
\usage{
dimred_plots_other_vars_fn(sce_dimred, dimred_plots_other_vars_params)
}
\arguments{
\item{sce_dimred}{(\emph{input target}) A \code{SingleCellExperiment} object. Input target in the integration plan is
\code{sce_int_final_clustering}.}

\item{dimred_plots_other_vars_params}{(\emph{input target}) A tibble.}
}
\value{
A tibble. \emph{Output target}: \code{dimred_plots_other_vars}
}
\description{
Make a dimred plot of selected variables.
}
\concept{sce_visualization}
