% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{.apply_config_patches}
\alias{.apply_config_patches}
\title{Apply YAML config patches to package-bundled configs.}
\usage{
.apply_config_patches(
  patches_dir,
  analysis_config_dir,
  pipeline_config_dir = getOption("scdrake_pipeline_config_dir")
)
}
\arguments{
\item{patches_dir}{A character scalar: path to directory with config patches. The files there must be named the same
as the bundled ones, e.g. \verb{02_norm_clustering.yaml}.}

\item{analysis_config_dir}{A character scalar: path to directory with analysis config files.}

\item{pipeline_config_dir}{A character scalar: path to directory with pipeline config file (\code{pipeline.yaml}).}
}
\value{
Invisibly \code{NULL}.
}
\description{
This is used in \code{tests/testthat/test-run_pipeline.R} tests where some modifications of default
config files are needed. The patches are located in \code{tests/testthat/run_pipeline_config_patches}.

Before the patches are applied, their local versions are updated from the default ones, similar to standard behaviour
(\code{\link[=update_config]{update_config()}}).
}
\details{
\code{pipeline_config_dir} directory will be used if \code{pipeline.default.yaml} file is present in \code{patches_dir}.
}
