% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_process_integration.R
\name{.check_integration_methods}
\alias{.check_integration_methods}
\title{Check for valid \code{INTEGRATION_SOURCES} parameter.}
\usage{
.check_integration_methods(
  integration_methods,
  method_names = c("uncorrected", "rescaling", "regression", "mnn", "harmony"),
  params = c("pca_selection_method", "pca_forced_pcs", "tsne_perp", "tsne_max_iter")
)
}
\arguments{
\item{integration_methods}{A named list.}

\item{method_names}{A character vector: valid names of integration methods.}

\item{params}{A character vector: names of valid parameters for each integration method.}
}
\value{
Invisibly \code{TRUE} if all checks are valid.
}
\description{
Check for valid \code{INTEGRATION_SOURCES} parameter.
}
\concept{internal}
