% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_process_common.R
\name{.check_marker_sources}
\alias{.check_marker_sources}
\title{Check for correct structure of \code{CLUSTER_MARKERS_SOURCES} or \code{CONTRASTS_SOURCES}.}
\usage{
.check_marker_sources(
  marker_sources,
  check_test_params = FALSE,
  marker_type = c("global", "contrast")
)
}
\arguments{
\item{marker_sources}{A named list.}

\item{check_test_params}{A logical scalar: if \code{TRUE}, check also \verb{params_*} for each test type.}

\item{marker_type}{A character scalar: \code{"global"} for cluster markers, \code{"contrast"} for contrasts.}
}
\value{
\code{TRUE} invisibly if all checks are valid.
}
\description{
Check for correct structure of \code{CLUSTER_MARKERS_SOURCES} or \code{CONTRASTS_SOURCES}.
}
