% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_process_common.R
\name{.get_dict_param}
\alias{.get_dict_param}
\title{Unlist a list of length one containing a named list.}
\usage{
.get_dict_param(
  param,
  not_empty = FALSE,
  empty_to_null = TRUE,
  name = deparse(substitute(param))
)
}
\arguments{
\item{param}{A list of length one containing a named list.}

\item{not_empty}{A logical scalar: if \code{TRUE}, throw error when \code{param} is \code{NULL} or empty list.}

\item{empty_to_null}{A logical scalar: if \code{TRUE} and \code{param} is empty list, return \code{NULL}.}

\item{name}{A passed \code{param} symbol, used for error messages.}
}
\value{
A named list, or \code{NULL} if \code{empty_to_null} is \code{TRUE} and the conditions are met.
}
\description{
This is used for some YAML parameters which need to be defined this way, otherwise they would get
overwritten by default values. Example of such parameter:

\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{PARAM:
  - var_1:
      a: 1
    var_2:
      b: 2
}\if{html}{\out{</div>}}
}
\concept{internal}
