% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_process_common.R
\name{.hereize_paths}
\alias{.hereize_paths}
\title{Using the \code{here} package, contruct paths relative to project's root directory.}
\usage{
.hereize_paths(cfg, paths)
}
\arguments{
\item{cfg}{A named list.}

\item{paths}{A character vector: names of items in \code{cfg} with paths to modify.}
}
\value{
A modified \code{cfg}.
}
\description{
Using the \code{here} package, contruct paths relative to project's root directory.
}
\concept{internal}
