% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_load.R
\name{.load_config_group}
\alias{.load_config_group}
\title{Load a group of YAML config files.}
\usage{
.load_config_group(dir, cfg_defs, cfg_pipeline = NULL, process = TRUE, ...)
}
\arguments{
\item{dir}{A character scalar: path to directory with YAML config files.}

\item{cfg_defs}{A list of lists of length three.
In each list, the first item is a name of YAML file in \code{dir}, the second one is a name in the resulting named list,
and the third one is a processing function.}

\item{cfg_pipeline}{A \code{scdrake_list} object: pipeline config (see \code{\link[=load_pipeline_config]{load_pipeline_config()}})
obtained from \code{pipeline.yaml} file located in pipeline config directory.
If \code{NULL}, it will be loaded using the path defined in \code{scdrake_pipeline_config_dir} option.}

\item{process}{A logical scalar: if \code{TRUE}, apply a processing function on its associated config list.}

\item{...}{Passed to \code{\link[=load_pipeline_config]{load_pipeline_config()}}, \code{\link[=load_config]{load_config()}}, and \code{\link[=.load_configs]{.load_configs()}}.}
}
\value{
A named \code{scdrake_list} of \code{scdrake_list}s for each entry in \code{cfg_defs} list.
}
\description{
Load a group of YAML config files.
}
