% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_load.R
\name{.load_configs}
\alias{.load_configs}
\title{Load a list of YAML config files.}
\usage{
.load_configs(dir, cfg_defs, other_variables = NULL, process = TRUE, ...)
}
\arguments{
\item{dir}{A path to directory with YAML config files.}

\item{cfg_defs}{A list of lists of length three.
In each list, the first item is a name of YAML file in \code{dir}, the second one is a name in the resulting named list,
and the third one is a processing function.}

\item{other_variables}{Passed to \code{load_config()}.}

\item{process}{A logical scalar: if \code{TRUE}, apply a processing function on its associated config list.}

\item{...}{Currently not used.}
}
\value{
A named list of lists of loaded YAML configs.
}
\description{
Load a list of YAML config files.
}
