% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_process_common.R
\name{.paths_to_base_dir}
\alias{.paths_to_base_dir}
\title{Construct paths relative to base directory.}
\usage{
.paths_to_base_dir(cfg, base_out_dir, paths)
}
\arguments{
\item{cfg}{A named list.}

\item{base_out_dir}{A character scalar: path to base directory.}

\item{paths}{A character vector: names of items in \code{cfg} with paths to modify.}
}
\value{
A modified \code{cfg}.
}
\description{
Construct paths relative to base directory.
}
\concept{internal}
