% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_update.R
\name{.update_configs_recursive}
\alias{.update_configs_recursive}
\title{Recursively update local configs using a file glob \verb{*.default.yaml}.}
\usage{
.update_configs_recursive(dir, ...)
}
\arguments{
\item{dir}{A character scalar: path to directory to recursively search for config files.}

\item{...}{Passed to \code{update_config()}.}
}
\description{
Recursively update local configs using a file glob \verb{*.default.yaml}.
}
\examples{
\dontrun{
.update_configs_recursive("config/single_sample")
}

}
\concept{internal}
