% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{download_example_data}
\alias{download_example_data}
\alias{download_pbmc1k}
\alias{download_pbmc3k}
\title{Download PBMC example data from 10x Genomics.}
\usage{
download_pbmc1k(out_dir, ask = TRUE, verbose = getOption("scdrake_verbose"))

download_pbmc3k(out_dir, ask = TRUE, verbose = getOption("scdrake_verbose"))
}
\arguments{
\item{out_dir}{A character scalar: path to directory to which will be data downloaded and extracted.
Will be automatically created if it does not exist.}

\item{ask}{A logical scalar: if \code{TRUE}, ask before download.}

\item{verbose}{A logical scalar: if \code{TRUE}, be verbose.
The default value is obtained from \code{getOption("scdrake_verbose")}.}
}
\value{
A vector of downloaded files (invisibly).
}
\description{
Those data are files for raw feature barcode matrices.
\itemize{
\item \code{download_pbmc1k()}:
\href{https://www.10xgenomics.com/resources/datasets/1-k-pbm-cs-from-a-healthy-donor-v-3-chemistry-3-standard-3-0-0}{PBMC 1k dataset}
(v3 chemistry, Cell Ranger 3.0.0)
\item \code{download_pbmc3k()}:
\href{https://www.10xgenomics.com/resources/datasets/3-k-pbm-cs-from-a-healthy-donor-1-standard-1-1-0}{PBMC 3k dataset}
(v2 chemistry, Cell Ranger 1.1.0)
}

\strong{All credits go to 10x Genomics.}
Visit \url{https://www.10xgenomics.com/resources/datasets} for more information.
}
\concept{download_example_data}
