% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yq.R
\name{download_yq}
\alias{download_yq}
\title{Download the binary of the \code{yq} tool (version 3.4.1).}
\usage{
download_yq(
  url = NULL,
  os = NULL,
  arch = NULL,
  yq_version = "3.4.1",
  destfile = get_yq_default_path(),
  set_as_default = TRUE,
  do_check = TRUE,
  ask = TRUE,
  overwrite = FALSE,
  dry = FALSE,
  verbose = getOption("scdrake_verbose")
)
}
\arguments{
\item{url}{A character scalar: URL from which binary will be downloaded.
If \code{NULL}, it will be created from \code{os}, \code{arch}, and \code{yq_version}.}

\item{os}{A character scalar: specifies the current OS (\code{"linux"}, \code{"windows"}, or \code{"darwin"} for macOS).
If \code{NULL}, it will be determined from \code{.Platform$OS.type}.}

\item{arch}{A character scalar: specifies the current CPU architecture (\code{32}, \code{64}, or \code{"arm64"}).
If \code{NULL}, it will be determined from \code{.Machine$sizeof.pointer} (\code{4 -> 32}, \code{8 -> 64}) or from \link{R.version} for
\code{"arm64"} (which is identified as \code{"aarch64"} for Apple's M1 chipsets).}

\item{yq_version}{A character scalar: \code{yq} tool version.}

\item{destfile}{A character scalar: destination file.}

\item{set_as_default}{A logical scalar: if \code{TRUE}, set \code{options(scdrake_yq_binary = destfile)}.}

\item{do_check}{A logical scalar: if \code{TRUE}, functionality of the downloaded (or present) \code{yq} binary will be checked.}

\item{ask}{A logical scalar: if \code{TRUE}, ask before \code{yq} binary is downloaded.}

\item{overwrite}{A logical scalar: if \code{TRUE}, overwrite the existing \code{yq} binary if exists.}

\item{dry}{A logical scalar: if \code{TRUE}, do not download the \code{yq} binary and just return its URL.}

\item{verbose}{A logical scalar: if \code{TRUE}, be verbose.
The default value is obtained from \code{getOption("scdrake_verbose")}.}
}
\value{
A character scalar of length two: \code{destfile} and URL to \code{yq} binary.
Note that on Windows platform, if \code{destfile} doesn't contain the \code{.exe} extension,
it will be automatically added.

If \code{!overwrite && fs::file_exists(destfile)}, then \code{NULL} invisibly.
}
\description{
This function has side-effects by default, see the \code{set_as_default} argument.
}
\concept{yq_tool}
