% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{filter_markers}
\alias{filter_markers}
\title{Filter a dataframe with cluster markers.}
\usage{
filter_markers(
  markers,
  top_n,
  top_n_wt = c("top", "fdr", "lfc", "auc"),
  distinct_by = NULL
)
}
\arguments{
\item{markers}{A dataframe-like object.}

\item{top_n}{A numeric scalar: number of top markers to keep.}

\item{top_n_wt}{A character scalar: name of column in \code{markers} by which top markers will be determined:
\itemize{
\item \code{"top"}: the minimum rank across all pairwise comparisons. Only possible if markers were computed
with \code{pval.type="any"}.
\item \code{"fdr"}: False Discovery Rate
\item \code{"lfc"}: log2 fold change
\item \code{"auc"}: Area Under Curve. Only possible if marker test was \code{"wilcox"}.
}}

\item{distinct_by}{A character scalar: name of column in \code{markers} by which unique rows will be kept.}
}
\value{
A filtered \code{markers} object.
}
\description{
Filter a dataframe with cluster markers.
}
\concept{sc_markers}
