% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_nulls}
\alias{filter_nulls}
\title{Return list with removed \code{NULL} elements.}
\usage{
filter_nulls(l)
}
\arguments{
\item{l}{A list.}
}
\value{
A list without \code{NULL} elements.
}
\description{
Return list with removed \code{NULL} elements.
}
\examples{
filter_nulls(list(a = 1, b = NULL, c = 3))
}
\concept{misc_utils}
