% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_text.R
\name{format_shell_command}
\alias{format_shell_command}
\title{Format a shell command as a Markdown codeblock.}
\usage{
format_shell_command(lines, language = "bash")
}
\arguments{
\item{lines}{A character vector: command lines.}

\item{language}{A character scalar: codeblock language.}
}
\value{
A character vector.
}
\description{
Format a shell command as a Markdown codeblock.
}
\concept{misc_text}
