% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{format_used_functions}
\alias{format_used_functions}
\title{Use the \code{{downlit}} package to generate a Markdown list of autolinked functions wrapped inside \verb{<details>} tags.}
\usage{
format_used_functions(functions, do_cat = FALSE)
}
\arguments{
\item{functions}{A character scalar: functions to be autolinked, e.g. \code{scran::clusterCells()}.}

\item{do_cat}{A logical scalar: if \code{TRUE}, cat the result to stdout.}
}
\value{
A character scalar. Invisibly if \code{do_cat} is \code{TRUE}.
}
\description{
Use the \code{{downlit}} package to generate a Markdown list of autolinked functions wrapped inside \verb{<details>} tags.
}
\examples{
format_used_functions(c("scran::clusterCells()", "Seurat::DimPlot()"), do_cat = TRUE)

}
\concept{misc_html}
