% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{generate_cell_annotation_plots_section}
\alias{generate_cell_annotation_plots_section}
\title{Generate a section with dimred plots used in some RMarkdown files.}
\usage{
generate_cell_annotation_plots_section(
  dimred_plots_cell_annotation_files,
  cell_annotation_diagnostic_plots,
  dimred_plots_rel_start = ".",
  cell_annotation_diagnostic_plots_rel_start = ".",
  main_header = "Cell annotation",
  main_header_level = 1,
  text = NULL
)
}
\arguments{
\item{dimred_plots_cell_annotation_files}{A \code{tibble}.}

\item{cell_annotation_diagnostic_plots}{A \code{tibble} or \code{NULL}.}

\item{dimred_plots_rel_start, cell_annotation_diagnostic_plots_rel_start}{A character scalar: relative start of path to directory with dimred plots, selected markers or cell annotation
diagnostic plots PDFs. See the \emph{Relative links} section in \code{\link[=create_a_link]{create_a_link()}}.}

\item{main_header}{A character scalar: text of the main header.}

\item{main_header_level}{A numeric scalar: level of the main header.}

\item{text}{A character text: additional text to be displayed on top of the section.}
}
\value{
Invisibly \code{NULL}.
}
\description{
The hierarchy is composed of tabsets:
\itemize{
\item Selected annotation reference
\itemize{
\item Dimred plots (UMAP, t-SNE, PCA)
}
}
}
\concept{misc_html}
