% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{generate_dimred_plots_section}
\alias{generate_dimred_plots_section}
\title{Generate a section with dimred plots used in some RMarkdown files.}
\usage{
generate_dimred_plots_section(
  dimred_plots_other_vars_files,
  selected_markers_plots_files,
  dimred_plots_rel_start = ".",
  selected_markers_files_rel_start = ".",
  main_header = "Dimensionality reduction",
  main_header_level = 1
)
}
\arguments{
\item{dimred_plots_other_vars_files}{A \code{tibble}.}

\item{selected_markers_plots_files}{A \code{tibble}.}

\item{dimred_plots_rel_start, selected_markers_files_rel_start}{A character scalar: relative start of path to directory with dimred plots or selected markers.
See the \emph{Relative links} section in \code{\link[=create_a_link]{create_a_link()}}.}

\item{main_header}{A character scalar: text of the main header.}

\item{main_header_level}{A numeric scalar: level of the main header.}
}
\value{
Invisibly \code{NULL}, Markdown elements are directly printed to stdout.
}
\description{
The hierarchy is composed of tabsets:
\itemize{
\item Dimred (UMAP, t-SNE, PCA)
\itemize{
\item Dimred plots of selected variables
}
}
}
\concept{misc_html}
