% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{generate_markers_results_section}
\alias{generate_markers_results_section}
\title{Generate a table with links to marker results.}
\usage{
generate_markers_results_section(df, group_var, dt_order_by, base_out_dir)
}
\arguments{
\item{df}{A dataframe-like object.}

\item{group_var}{A character scalar: name of grouping variable for \code{df}.}

\item{dt_order_by}{A character scalar: name of column in \code{df} to order rows by.}

\item{base_out_dir}{A character scalar: path to base out dir for \code{cluster_markers} or \code{contrasts} stage.
Used to determine relative paths from rendered HTML files to PDF files (marker plots, heatmaps).}
}
\value{
Invisibly \code{NULL}.
}
\description{
This function is used directly in Rmd files for its side effects (in chunks with \code{results = "asis"}).
}
\concept{sc_markers}
