% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{generate_stage_report}
\alias{generate_stage_report}
\title{Render RMarkdown document for a stage of the pipeline.}
\usage{
generate_stage_report(
  rmd_file,
  out_html_file_name,
  output_dir = fs::path_dir(out_html_file_name),
  css_file = here("Rmd/common/stylesheet.css"),
  params = list(),
  message = TRUE,
  warning = TRUE,
  echo = TRUE,
  quiet = TRUE,
  other_deps = NULL,
  drake_cache_dir = here(".drake"),
  ...
)
}
\arguments{
\item{rmd_file}{A character scalar: path to Rmd file to render.}

\item{out_html_file_name}{A character scalar: name of the output HTML file.}

\item{output_dir}{A character scalar: path to directory in which \code{out_html_file_name} will be created.
This directory will be created if it does not exist.}

\item{css_file}{A character scalar: path to CSS file which will be included in the resulting HTML.}

\item{params}{A list of RMarkdown document parameters:
passed to \code{\link[rmarkdown:render]{rmarkdown::render()}} together with \code{css_file} and \code{drake_cache_dir} items.}

\item{message, warning, echo}{A logical scalar: passed to \code{knitr::opts_chunk$set()}.}

\item{quiet}{A logical scalar: if \code{TRUE}, do not be verbose during rendering. Passed to \code{\link[rmarkdown:render]{rmarkdown::render()}}.}

\item{other_deps}{A list of symbols (unquoted): used inside \code{\link[drake:plan]{drake::plan()}} to mark other target dependencies,
which are not loaded (\code{\link[drake:readd]{drake::loadd()}}) or read (\code{\link[drake:readd]{drake::loadd()}}) inside the \code{rmd_file}.}

\item{drake_cache_dir}{A character scalar: path to \code{drake} cache directory.}

\item{...}{Passed to \code{\link[rmarkdown:render]{rmarkdown::render()}}.}
}
\value{
Invisibly \code{NULL}.
}
\description{
Render RMarkdown document for a stage of the pipeline.
}
\details{
If you are currently working within a \code{scdrake} project, you can review the Rmd files in the \verb{Rmd/} directory.

This is a list of Rmd files currently bundled with the \code{scdrake} package:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{fs::dir_tree(system.file("Rmd", package = "scdrake"))
#> /home/rstudio/scdrake_source/inst/Rmd
#> +-- common
#> |   +-- _footer.Rmd
#> |   +-- _header.Rmd
#> |   +-- cluster_markers.Rmd
#> |   +-- cluster_markers_table_template.Rmd
#> |   +-- clustering
#> |   |   +-- cluster_graph_leiden.Rmd
#> |   |   +-- cluster_graph_louvain.Rmd
#> |   |   +-- cluster_graph_walktrap.Rmd
#> |   |   +-- cluster_kmeans.Rmd
#> |   |   +-- cluster_sc3.Rmd
#> |   |   \\-- clustering.Rmd
#> |   +-- contrasts.Rmd
#> |   +-- contrasts_table_template.Rmd
#> |   \\-- stylesheet.css
#> +-- integration
#> |   +-- 01_integration.Rmd
#> |   \\-- 02_int_clustering.Rmd
#> \\-- single_sample
#>     +-- 01_input_qc.Rmd
#>     +-- 01_input_qc_children
#>     |   +-- cell_filtering_custom.Rmd
#>     |   +-- cell_filtering_qc.Rmd
#>     |   +-- empty_droplets.Rmd
#>     |   +-- gene_filtering_custom.Rmd
#>     |   \\-- gene_filtering_qc.Rmd
#>     +-- 02_norm_clustering.Rmd
#>     \\-- 02_norm_clustering_simple.Rmd
}\if{html}{\out{</div>}}

A short description of these Rmd files:
\itemize{
\item Common to single-sample and integration analyses:
\itemize{
\item Cluster markers: \code{common/cluster_markers.Rmd} (signpost),
\code{common/cluster_markers_table_template.Rmd} (actual table with markers)
\item Contrasts: \code{common/cluster_markers.Rmd}, \code{common/cluster_markers_table_template.Rmd}
}
\item Single-sample analysis:
\itemize{
\item Input QC: \verb{single_sample/01_input_qc.Rmd}.
Input data overview, empty droplets detection, cell and gene filtering visualization.
\item Normalization and clustering: \verb{single_sample/02_norm_clustering.Rmd},
\verb{single_sample/02_norm_clustering_simple.Rmd} (only clustering visualization - dimred plots).
Overview of normalization and highly variable genes selection, doublet score plots, selection of principal components,
clustering diagnostics, dimred plots of clusterings and other variables.
}
\item Integration analysis:
\itemize{
\item Integration: \verb{integration/01_integration.Rmd}.
Overview of input datasets, highly variable genes selection, selection of principal components, dimred plots
of integration, integration diagnostics.
\item Clustering: \verb{integration/02_int_clustering.Rmd}.
Clustering diagnostics, dimred plots of clusterings and other variables.
}
}

All Rmd files are also accompanied with a list of config parameters and runtime information.
}
\section{Working with Rmd files outside of a plan}{

Each Rmd file for a stage can be rendered (with \code{generate_stage_report()}, not \code{rmarkdown::render()} - see below), or
run interactively in RStudio, both outside of a \code{drake} plan. However, this assumes:
\itemize{
\item Your working directory is in the root of a \code{scdrake} project.
\item For interactive usage in RStudio, set \verb{Options -> R Markdown -> Basic -> Evaluate chunks in directory} to \code{"Project"}.
Otherwise the working directory will be in the location of a Rmd file while running R code chunks.
Also, all Rmd files are parametrized, i.e. they are referencing a \code{params} list, which is passed to
\code{rmarkdown::render()} by \code{generate_stage_report()}. In order for code chunks to work properly,
you have to create a variable named \code{params_} (it must be available in the global environment) containing a named list.
Currently, this list must contain only a single item: \code{drake_cache_dir} with path to \code{drake} cache which will be
used to load targets inside the Rmd files. Example: \code{params_ <- list(drake_cache_dir = ".drake_my_analysis")}
}

NOTE: It is not recommended to call \code{rmarkdown::render()} directly on a stage's Rmd file,
because \code{generate_stage_report()} also sets some temporary options and passes a proper \code{params} parameter.
}

\examples{
\dontrun{
generate_stage_report(
  rmd_file = here("Rmd/single_sample/01_input_qc.Rmd"),
  out_html_file_name = "01_input_qc.html",
  output_dir = here("my_output"),
  drake_cache_dir = here(".drake_cache")
)
}

}
\concept{misc_html}
