% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans_common_cell_annotation.R
\name{get_cell_annotation_subplan}
\alias{get_cell_annotation_subplan}
\title{Get a subplan for cell annotation via \code{SingleR}}
\usage{
get_cell_annotation_subplan(
  sce_target_name,
  sce_dimred_plots_target_name,
  cell_annotation_sources,
  cell_annotation_out_dir,
  report_dimred_names,
  dimred_plots_out_dir,
  do_heatmaps_
)
}
\arguments{
\item{sce_target_name, sce_dimred_plots_target_name}{A character scalar: name of target representing a SCE object that will be used for cell annotation or dimred plots.}

\item{cell_annotation_sources}{A tibble with cell annotation parameters.}

\item{cell_annotation_out_dir}{A character scalar: output directory for diagnostic plots.}

\item{report_dimred_names}{A character vector: dimred names for plots.}

\item{dimred_plots_out_dir}{A character scalar: output directory for dimred plots.}

\item{do_heatmaps_}{A logical scalar: if \code{TRUE}, make heatmaps annotated by predicted cell types and cluster memberships.}
}
\value{
\code{\link[drake:drake_plan]{drake::drake_plan()}}
}
\description{
Get a subplan for cell annotation via \code{SingleR}
}
\concept{get_subplan_cell_annotation}
