% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans_common_clustering.R
\name{get_clustering_graph_subplan}
\alias{get_clustering_graph_subplan}
\title{Get a subplan for graph-based clustering.}
\usage{
get_clustering_graph_subplan(
  sce_target_name,
  dimred,
  cluster_graph_louvain_enabled,
  cluster_graph_louvain_resolutions_,
  cluster_graph_walktrap_enabled,
  cluster_graph_leiden_enabled,
  cluster_graph_leiden_resolutions_,
  cluster_graph_snn_k_,
  cluster_graph_snn_type_,
  is_integration,
  plots_out_dir
)
}
\arguments{
\item{sce_target_name}{A character scalar: name of target representing a SCE object that will be used for SNN graph construction.}

\item{dimred}{A character scalar: which \code{reducedDim()} to use for clustering.}

\item{cluster_graph_louvain_resolutions_, cluster_graph_leiden_resolutions_}{A numeric vector: resolutions to calculate for the
selected clustering method.}

\item{cluster_graph_leiden_enabled, cluster_graph_louvain_enabled, cluster_graph_walktrap_enabled}{A logical scalar: if \code{FALSE}, disable the selected clustering and set all corresponding targets to \code{NULL}.}

\item{cluster_graph_snn_k_, cluster_graph_snn_type_}{Passed to \code{\link[=graph_snn_fn]{graph_snn_fn()}}.}

\item{is_integration}{A logical scalar: if \code{TRUE}, clustering results will be named as \verb{cluster_int_graph_<method>[_r<r>]}.
If \code{FALSE}, they will be named as \verb{cluster_graph_<method>[_r<r>]}.}

\item{plots_out_dir}{A character scalar: path to output directory to save plots (e.g. \code{\link[clustree:clustree]{clustree::clustree()}}).}
}
\value{
\code{\link[drake:drake_plan]{drake::drake_plan()}}
}
\description{
Currently supported community-based detection algorithms are:
\itemize{
\item \code{\link[igraph:cluster_leiden]{igraph::cluster_leiden()}}
\item \code{\link[igraph:cluster_louvain]{igraph::cluster_louvain()}}
\item \code{\link[igraph:cluster_walktrap]{igraph::cluster_walktrap()}}
}

See \code{\link[=run_graph_based_clustering]{run_graph_based_clustering()}} for more details.
}
\concept{get_subplan_clustering}
