% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans_common_clustering.R
\name{get_clustering_kmeans_subplan}
\alias{get_clustering_kmeans_subplan}
\title{Get a subplan for k-means clustering.}
\usage{
get_clustering_kmeans_subplan(
  sce_target_name,
  dimred,
  cluster_kmeans_k_enabled,
  cluster_kmeans_k,
  cluster_kmeans_kbest_enabled,
  is_integration,
  plots_out_dir
)
}
\arguments{
\item{sce_target_name}{A character scalar: name of target representing a SCE object that will be used for k-means clustering.}

\item{dimred}{A character scalar: which \code{reducedDim()} to use for clustering.}

\item{cluster_kmeans_k_enabled, cluster_kmeans_kbest_enabled}{A logical scalar: if \code{FALSE}, disable the selected clustering and set all corresponding targets to \code{NULL}.}

\item{cluster_kmeans_k}{An integer vector: \code{k}s for k-means.}

\item{is_integration}{A logical scalar: if \code{TRUE}, clustering results will be named as \verb{cluster_int_kmeans_<method>[_k<k>]}.
If \code{FALSE}, they will be named as \verb{cluster_kmeans_<method>[_k<k>]}.}

\item{plots_out_dir}{A character scalar: path to output directory to save plots (e.g. \code{\link[clustree:clustree]{clustree::clustree()}}).}
}
\value{
\code{\link[drake:drake_plan]{drake::drake_plan()}}
}
\description{
Besides k-means for selected number of \code{k}s, there is also a best \code{k} algorithm.
}
\concept{get_subplan_clustering}
