% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans_common_clustering.R
\name{get_clustering_sc3_subplan}
\alias{get_clustering_sc3_subplan}
\title{Get a subplan for SC3 clustering.}
\usage{
get_clustering_sc3_subplan(
  sce_target_name,
  cluster_sc3_enabled,
  cluster_sc3_k_,
  cluster_sc3_n_cores,
  is_integration,
  plots_out_dir,
  seed = 1
)
}
\arguments{
\item{sce_target_name}{A character scalar: name of target representing a SCE object that will be used for SC3 clustering.}

\item{cluster_sc3_enabled}{A logical scalar: if \code{FALSE}, disable the selected clustering and set all corresponding targets to \code{NULL}.}

\item{cluster_sc3_k_}{An integer vector: \code{k}s for SC3.}

\item{cluster_sc3_n_cores}{An integer scalar: number of cores to use for parallel execution of the SC3 algorithm.}

\item{is_integration}{A logical scalar: if \code{TRUE}, clustering results will be named as \verb{cluster_int_sc3_k<k>}.
If \code{FALSE}, they will be named as \verb{cluster_sc3_k<k>}.}

\item{plots_out_dir}{A character scalar: path to output directory to save plots (e.g. \code{\link[clustree:clustree]{clustree::clustree()}}).}

\item{seed}{An integer scalar: random seed for SC3.}
}
\value{
\code{\link[drake:drake_plan]{drake::drake_plan()}}
}
\description{
Get a subplan for SC3 clustering.
}
\concept{get_subplan_clustering}
