% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans_common_clustering.R
\name{get_clustering_subplan}
\alias{get_clustering_subplan}
\title{Get subplan for clustering.}
\usage{
get_clustering_subplan(
  cfg,
  sce_clustering_target_name,
  sce_dimred_plots_target_name,
  dimred,
  report_dimred_names,
  dimred_plots_out_dir,
  other_plots_out_dir,
  is_integration,
  seed = 1
)
}
\arguments{
\item{cfg}{A list of parameters for \verb{02_norm_clustering} or \verb{02_int_clustering} stage.}

\item{sce_clustering_target_name, sce_dimred_plots_target_name}{A character scalar: name of target representing a SCE object that will be used for SC3 clustering and dimred plots, respectively.}

\item{dimred}{A character scalar: which \code{reducedDim()} to use for clustering. This is only applied to graph-based and k-means clustering.}

\item{report_dimred_names}{A character vector: dimreds to use for plotting clustering results.}

\item{dimred_plots_out_dir, other_plots_out_dir}{A character scalar: path to output directory to save plots.}

\item{is_integration}{A logical scalar: if \code{TRUE}, clustering results will be named with \verb{cluster_int_*} prefix.}

\item{seed}{An integer scalar: random seed for SC3.}
}
\value{
A combined \code{\link[drake:drake_plan]{drake::drake_plan()}} from:
\itemize{
\item \code{\link[=get_clustering_graph_subplan]{get_clustering_graph_subplan()}}
\item \code{\link[=get_clustering_kmeans_subplan]{get_clustering_kmeans_subplan()}}
\item \code{\link[=get_clustering_sc3_subplan]{get_clustering_sc3_subplan()}}
}
}
\description{
Besides clustering itself, this plan also includes dimred plots of computed clusters.
}
\concept{get_subplan_clustering}
