% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans_common.R
\name{get_dimred_plots_other_vars_subplan}
\alias{get_dimred_plots_other_vars_subplan}
\title{Get a subplan for dimensionality reduction plots of selected variables.}
\usage{
get_dimred_plots_other_vars_subplan(
  sce_target_name,
  report_dimred_plots_other,
  report_dimred_names,
  dimred_plots_out_dir
)
}
\arguments{
\item{sce_target_name}{A character scalar: name of target that stores SCE object, which will be used for plotting.}

\item{report_dimred_plots_other}{A character vector: variables in \verb{colData(<sce_target_name>)} to plot.}

\item{report_dimred_names}{A character vector: dimred names for plotting (\code{"pca"}, \code{"tsne"}, \code{"umap"}).}

\item{dimred_plots_out_dir}{A character scalar: output directory to save PDFs into.}
}
\value{
\code{\link[drake:drake_plan]{drake::drake_plan()}}
}
\description{
Get a subplan for dimensionality reduction plots of selected variables.
}
\concept{get_subplan_common}
