% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_sample_input_qc.R
\name{get_gene_filter}
\alias{get_gene_filter}
\title{Get a logical filter for ratio of cells expressing a gene and minimum number of UMI per gene.}
\usage{
get_gene_filter(sce, min_ratio_cells, min_umi)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object.}

\item{min_ratio_cells}{A numeric scalar: minimum ratio of cells expressing a gene.}

\item{min_umi}{A numeric scalar: minimum number of UMI per gene across all cells.}
}
\value{
A logical vector.
}
\description{
A gene is considered expressed when number of its UMIs across all cells is greater than \code{min_umi} and
at the same time it is expressed in at least \code{min_ratio_cells} ratio of cells.
}
\concept{single_sample_input_qc_fn}
