% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_integration.R
\name{get_int_method_description}
\alias{get_int_method_description}
\title{Return description for an integration method.}
\usage{
get_int_method_description(
  int_method_name = c("uncorrected", "rescaling", "regression", "mnn", "harmony")
)
}
\arguments{
\item{int_method_name}{A character scalar: integration method name.}
}
\value{
A named list with character scalars: \code{header}, \code{description}, \code{fn_link}.
}
\description{
Return description for an integration method.
}
\concept{integration_integration_fn}
