% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{get_legend_35}
\alias{get_legend_35}
\title{Extract legend from plot}
\usage{
get_legend_35(plot)
}
\arguments{
\item{`ggplot2`}{object as input parameter}
}
\description{
Extract legend and add it to the selected place of composite image
}
\concept{sce_visualization}
