% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_pca_gene_var_pcs}
\alias{get_pca_gene_var_pcs}
\title{Get a number of PCs corresponding to biological variation.}
\usage{
get_pca_gene_var_pcs(
  sce_pca,
  BSPARAM = BiocSingular::IrlbaParam(),
  BPPARAM = BiocParallel::SerialParam()
)
}
\arguments{
\item{sce_pca}{A \code{SingleCellExperiment} object with calculated PCA.}

\item{BSPARAM}{A \link[BiocSingular:BiocSingularParam]{BiocSingular::BiocSingularParam} object.}

\item{BPPARAM}{A \link[BiocParallel:BiocParallelParam-class]{BiocParallel::BiocParallelParam} object.}
}
\value{
An integer scalar.
}
\description{
This strategy is further described in
\href{https://bioconductor.org/books/3.15/OSCA.advanced/dimensionality-reduction-redux.html#using-the-technical-noise}{OSCA}.
}
\concept{sc_clustering}
