% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_pca_selected_pcs}
\alias{get_pca_selected_pcs}
\title{Get a final selection of number of PCs.}
\usage{
get_pca_selected_pcs(
  sce_pca,
  pca_elbow_pcs,
  pca_gene_var_pcs,
  pca_selection_method = c("elbow", "technical_noise", "forced"),
  pca_forced_pcs = NULL
)
}
\arguments{
\item{sce_pca}{A \code{SingleCellExperiment} object with calculated PCA.}

\item{pca_elbow_pcs}{A numeric scalar: number of PCs selected by elbow point strategy.}

\item{pca_gene_var_pcs}{A numeric scalar: number of PCs selected by technical noise strategy.}

\item{pca_selection_method}{A character scalar: final selection strategy.}

\item{pca_forced_pcs}{A numeric scalar: constant number of PCs defined in config file.}
}
\value{
A modified \code{SingleCellExperiment} object:
\itemize{
\item The full PCA matrix in \code{"pca_all"} slot of \code{reducedDim()}.
\item The subsetted PCA matrix in \code{"pca"} slot of \code{reducedDim()}.
\item New items added to \code{metadata()}: \code{pca_selection_method}, \code{pca_selected_pcs}
}
}
\description{
The number of PCs and the selection strategy is controlled:
\itemize{
\item For single-sample analysis, via \code{PCA_SELECTION_METHOD} and \code{PCA_FORCED_PCS} parameters in \verb{02_norm_clustering.yaml}.
\item For integration analysis, via \code{pca_selection_method} and \code{pca_forced_pcs} subparameters of \code{INTEGRATION_METHODS}
parameter in \verb{01_integration.yaml}.
}
}
\concept{sc_clustering}
