% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_text.R
\name{get_random_strings}
\alias{get_random_strings}
\title{Generate random strings compounded from alphabetical characters by default.}
\usage{
get_random_strings(n, length, pattern = "[A-Za-z]")
}
\arguments{
\item{n, length, pattern}{Passed to \code{stringi::stri_rand_strings()}.}
}
\value{
A character vector.
}
\description{
This is a wrapper around \code{stringi::stri_rand_strings()}.
}
\concept{misc_text}
