% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sys_env}
\alias{get_sys_env}
\title{Read an environment variable and if needed, assign a default value.}
\usage{
get_sys_env(
  var,
  default = NULL,
  type = c("character", "double", "integer", "logical"),
  verbose = getOption("scdrake_verbose")
)
}
\arguments{
\item{var}{A character scalar: name of environment variable.}

\item{default}{A default value (different from \code{NULL}) when environment variable is unset (empty string).}

\item{type}{A character scalar: to which R type will be the variable coerced.}

\item{verbose}{A logical scalar: if \code{TRUE}, be verbose.
The default value is obtained from \code{getOption("scdrake_verbose")}.}
}
\value{
A character scalar if the environment variable is set, \code{default} if not, or empty character scalar when
\code{default = NULL} and the variable is unset.
}
\description{
Read an environment variable and if needed, assign a default value.
}
\examples{
get_sys_env("PATH")
get_sys_env("NON_EXISTING")
get_sys_env("NON_EXISTING", default = "default_value")
}
\concept{misc_utils}
