% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tmp_dir}
\alias{get_tmp_dir}
\title{Get a path to the temporary directory according to the OS.}
\usage{
get_tmp_dir()
}
\description{
Source: \url{https://stackoverflow.com/questions/16474696/read-system-tmp-dir-in-r}
}
\concept{misc_utils}
