% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hvg.R
\name{get_top_hvgs}
\alias{get_top_hvgs}
\title{Get top N highly variable genes (HVGs) by a specified metric.}
\usage{
get_top_hvgs(
  sce_norm,
  hvg_metric_fit,
  hvg_selection_value,
  hvg_metric = c("gene_var", "gene_cv2", "sctransform"),
  hvg_selection = c("top", "significance", "threshold")
)
}
\arguments{
\item{sce_norm}{A \code{SingleCellExperiment} object with computed HVG metric.}

\item{hvg_metric_fit}{A \code{DataFrame} with HVG metric fit as produced by e.g. \code{\link[scran:modelGeneVar]{scran::modelGeneVar()}}.}

\item{hvg_selection_value}{A numeric scalar: threshold value to select HVGs. This depends on \code{hvg_metric}.}

\item{hvg_metric}{A character scalar: type of HVG metric.
If \code{sctransform} is used, HVGs are selected by the underlying method, and number of them is controlled by
\code{SCT_N_HVG} parameter in \verb{02_norm_clustering.yaml}.
For the other metric types, see the \code{hvg_selection} and \code{hvg_selection_value} parameters.}

\item{hvg_selection}{A character scalar: method to use for selection of HVGs.
This is only relevant when \code{hvg_metric} is "gene_var" or "gene_cv2".
See \url{http://bioconductor.org/books/release/OSCA/feature-selection.html#hvg-selection} for more details.
\itemize{
\item \code{"top"}: Take top X genes according to a metric.
\code{"bio"} and \code{"ratio"} columns are used for \code{hvg_metric} \code{"gene_var"} and \code{"gene_cv2"}, respectively.
\item \code{"significance"}: Use FDR threshold.
\item \code{"threshold"}: Use threshold on the minimum value of a metric.
\code{"bio"} and \code{"ratio"} columns are used for \code{hvg_metric} of \code{"gene_var"} and \code{"gene_cv2"}, respectively.
}}
}
\value{
A character vector of HVG IDs (ENSEMBL).
}
\description{
Types of HVG metrics are specified in \verb{02_norm_clustering.yaml} and \verb{01_integration.yaml} configs.
See the corresponding section there.
}
\concept{sc_hvg}
