% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_sample_input_qc.R
\name{get_used_qc_filters_operator_desc}
\alias{get_used_qc_filters_operator_desc}
\title{Return an informative message about the used operator to join cell QC filters}
\usage{
get_used_qc_filters_operator_desc(operator = c("&", "|"))
}
\arguments{
\item{operator}{A character scalar: used operator.}
}
\value{
A character scalar: the message
}
\description{
Return an informative message about the used operator to join cell QC filters
}
\concept{single_sample_input_qc_fn}
