% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yq.R
\name{get_yq_default_path}
\alias{get_yq_default_path}
\title{Return a default download path for the \code{yq} tool's binary.}
\usage{
get_yq_default_path()
}
\value{
A character scalar.
}
\description{
The difference between platforms is that on Windows the \code{.exe} extension is added.
}
