% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visualization.R
\name{getDistinctColors}
\alias{getDistinctColors}
\title{Generating distinct colors}
\usage{
getDistinctColors(n)
}
\arguments{
\item{n}{Number of unique factors in data}
}
\description{
Generating distinct colors for rendering color layer for factor variable
}
\concept{spatial_visualization}
