% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%&&\%}
\alias{\%&&\%}
\title{A short-circuit evaluation-like function similar to \code{bash}'s \code{&&} operator.}
\usage{
x \%&&\% y
}
\arguments{
\item{x}{A logical scalar: left-hand side argument.}

\item{y}{A right-hand side expression.}
}
\value{
\code{y} if \code{x} is \code{TRUE}, else \code{invisible(NULL)}
}
\description{
The right-hand side (\code{y}) is only evaluated if the left-hand side (\code{x}) is identical to \code{TRUE}.
Useful for verbosity (\code{verbose \%&&\% message("Hello world!")}).
}
\examples{
TRUE \%&&\% message("printed")
FALSE \%&&\% message("not printed")
}
\concept{misc_utils}
