% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{highlight_points}
\alias{highlight_points}
\title{Highlight points belonging to certain levels.}
\usage{
highlight_points(p, column_name, levels, alpha_val = 0.1)
}
\arguments{
\item{p}{A \code{ggplot2} object.}

\item{column_name}{A character scalar: name of a column in \code{p} from which levels will be read.}

\item{levels}{A character vector: levels to highlight.}

\item{alpha_val}{A numeric scalar: alpha value to set for levels other than those in \code{levels}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Highlighting means that alpha value of other levels will be decreased.
}
\examples{
p <- ggplot2::ggplot(
  mtcars,
  ggplot2::aes(x = cyl, y = mpg, color = factor(am))
) +
  ggplot2::geom_point()
highlight_points(p, "am", "0", alpha_val = 0.25)

}
\concept{sce_visualization}
