% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_integration.R
\name{hvg_int_list_fn}
\alias{hvg_int_list_fn}
\title{Get a named list of HVG data (gene IDs, fit, metadata).}
\usage{
hvg_int_list_fn(
  sce_int_multibatchnorm,
  hvg_selection_value,
  hvg_combination = c("hvg_metric", "intersection", "union", "all"),
  hvg_selection = c("top", "significance", "threshold")
)
}
\arguments{
\item{sce_int_multibatchnorm}{(\emph{input target}) A \code{SingleCellExperiment} object
processed by \code{\link[batchelor:multiBatchNorm]{batchelor::multiBatchNorm()}}.}

\item{hvg_selection_value, hvg_combination, hvg_selection}{See \code{\link[=sce_int_combine_hvgs]{sce_int_combine_hvgs()}}.}
}
\value{
A named list. \emph{Output target}: \code{hvg_int_list}
}
\description{
Get a named list of HVG data (gene IDs, fit, metadata).
}
\concept{integration_integration_fn}
