% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_integration.R
\name{hvg_plot_int_fn}
\alias{hvg_plot_int_fn}
\title{Make a HVG plot for uncorrected \code{SingleCellExperiment} object (processed by \code{\link[batchelor:multiBatchNorm]{batchelor::multiBatchNorm()}}).}
\usage{
hvg_plot_int_fn(sce_int_uncorrected, ...)
}
\arguments{
\item{sce_int_uncorrected}{(\emph{input target}) A tibble.}

\item{...}{Passed to \code{\link[=plot_hvg]{plot_hvg()}}.}
}
\value{
A modified \code{sce_int_uncorrected} with appended HVG plot (\code{ggplot2} object).
\emph{Output target}: \code{hvg_plot_int}
}
\description{
Make a HVG plot for uncorrected \code{SingleCellExperiment} object (processed by \code{\link[batchelor:multiBatchNorm]{batchelor::multiBatchNorm()}}).
}
\concept{integration_integration_fn}
