% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{init_project}
\alias{init_project}
\title{Initialize a new \code{scdrake} project.}
\usage{
init_project(
  dir = NULL,
  use_rstudio = TRUE,
  set_active_project = TRUE,
  set_wd = TRUE,
  download_example_data = FALSE,
  ask = TRUE,
  verbose = getOption("scdrake_verbose"),
  ...
)
}
\arguments{
\item{dir}{A character scalar: path to directory in which the project will be created.
If \code{NULL}, the current working directory will be used. Subdirectories will be created recursively.}

\item{use_rstudio}{A logical scalar: if \code{TRUE}, initiate an RStudio project inside \code{dir}.}

\item{set_active_project}{A logical scalar: if \code{TRUE}, set active project to \code{dir}.}

\item{set_wd}{A logical scalar: if \code{TRUE}, change the current working directory to \code{dir}.}

\item{download_example_data}{A logical scalar: if \code{TRUE}, download example PBMC 1k and 3k data to \code{{dir}/example_data}.
See \code{\link[=download_pbmc1k]{download_pbmc1k()}} and \code{\link[=download_pbmc3k]{download_pbmc3k()}} for more details on these datasets.}

\item{ask}{A logical scalar: if \code{TRUE}, do not prompt and overwrite an existing project.}

\item{verbose}{A logical scalar: if \code{TRUE}, be verbose.
The default value is obtained from \code{getOption("scdrake_verbose")}.}

\item{...}{Passed to \code{download_yq()}.}
}
\description{
Initialize a new \code{scdrake} project.
}
\details{
This function will:
\itemize{
\item Create the project's root directory.
\item (Optional) Switch the current working directory to the project's root and call \code{\link[here:i_am]{here::i_am()}}.
\item Copy config, RMarkdown, and \verb{_drake_*.R} files.
\item Create an empty \code{.here} file. This ensures that the \href{https://here.r-lib.org/}{here} package will be able to find
the project's root directory in case an RStudio project is not initialized.
\item Check whether the \code{yq} tool is available in \code{PATH} environment variable.
\item (Optional) Download the example data using \code{\link[=download_pbmc1k]{download_pbmc1k()}} and \code{\link[=download_pbmc3k]{download_pbmc3k()}}.
\item (Optional) Initialize an RStudio project and set it as the active project.
}
}
\examples{
\dontrun{
init_project("my_project")
}

}
