% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_integration.R
\name{int_diagnostics_df_fn}
\alias{int_diagnostics_df_fn}
\title{Compute and make integration diagnostics and plots for each integration method.}
\usage{
int_diagnostics_df_fn(sce_int_clustering_df, sce_int_raw_snn_clustering)
}
\arguments{
\item{sce_int_clustering_df}{(\emph{input target}) A tibble.}

\item{sce_int_raw_snn_clustering}{(\emph{input target}) A tibble.}
}
\value{
A tibble. \emph{Output target}: \code{int_diagnostics_df}
}
\description{
Compute and make integration diagnostics and plots for each integration method.
}
\concept{integration_integration_fn}
