% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_integration.R
\name{integration_methods_df_fn}
\alias{integration_methods_df_fn}
\title{Make a tibble with integration methods and their parameters.}
\usage{
integration_methods_df_fn(integration_methods, hvg_int, hvg_int_with_cc = NULL)
}
\arguments{
\item{integration_methods}{A named list of integration methods.
See the \code{INTEGRATION_METHODS} parameter in \verb{01_integration.yaml} config for more details.}

\item{hvg_int}{(\emph{input target}) A named list of HVG data for each \code{SingleCellExperiment} object.
Depending on settings, it may have CC-related HVGs removed.}

\item{hvg_int_with_cc}{(\emph{input target}) \code{NULL} or a named list of HVG data for each \code{SingleCellExperiment} object.
Depending on settings, it may have all genes (including CC-related) included.}
}
\value{
A \code{tibble}. \emph{Output target}: \code{integration_methods_df}
}
\description{
Make a tibble with integration methods and their parameters.
}
\concept{integration_integration_fn}
