% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lapply_rows}
\alias{lapply_rows}
\title{Apply a function over rows of a \code{data.frame}-like object and concatenate the results back to \code{tibble} or \code{data.frame}.}
\usage{
lapply_rows(df, as_scdrake_list = TRUE, return_tibble = TRUE, FUN, ...)
}
\arguments{
\item{df}{A \code{data.frame}-like object.}

\item{as_scdrake_list}{A logical scalar: if \code{TRUE}, each row-list in \code{FUN} will be converted to \code{\link[=scdrake_list]{scdrake_list()}}.}

\item{return_tibble}{If \code{TRUE}, a \code{tibble} with concatenated results is returned.
Otherwise results are coerced to \code{data.frame} and original rownames are set.}

\item{FUN}{A function to apply over \code{df} rows.}

\item{...}{Additional arguments passed to \code{FUN}.}
}
\value{
\code{tibble} or \code{data.frame} according to the \code{return_tibble} parameter.
}
\description{
Inside a \code{FUN}, a row of \code{df} will be available as a list or \code{\link[=scdrake_list]{scdrake_list()}} (default).
The \code{FUN} must return a named list for a proper concatenation of the results.
The named list can also contain elements of a length other than one, which are then wrapped in \code{list()}.
}
\examples{
library(magrittr)

fn <- function(row_list) {
  row_list$cyl_2 <- row_list$cyl**2
  row_list$colors <- c("red", "green", "blue")
  row_list$sublist <- mtcars[1:5, 1:5]
  return(row_list)
}

df <- lapply_rows(mtcars, FUN = fn)
head(df)

df2 <- lapply_rows(mtcars, return_tibble = FALSE, FUN = fn)
head(df2)
}
\concept{misc_utils}
