% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_names_to_values}
\alias{list_names_to_values}
\title{Append names of a list of lists as values in each sublist.}
\usage{
list_names_to_values(l, key = "name")
}
\arguments{
\item{l}{A named list of lists.}

\item{key}{A character scalar: name of item under which names of outer list will be appended.}
}
\value{
A named list of lists.
}
\description{
Append names of a list of lists as values in each sublist.
}
\examples{
list_names_to_values(list(a = list(1), b = list(c = 3)))

}
\concept{misc_utils}
