% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lists_to_tibble}
\alias{lists_to_tibble}
\title{Convert list of lists to tibble.}
\usage{
lists_to_tibble(l, do_replace_list_nulls = TRUE, do_bind_rows = TRUE)
}
\arguments{
\item{l}{A list of lists.}

\item{do_replace_list_nulls}{A logical scalar: if \code{TRUE}, first replace \code{NULL} items in each sublist by \code{NA}s.}

\item{do_bind_rows}{A logical scalar: if \code{TRUE}, bind all tibbles by rows, otherwise return a list of tibbles.}
}
\value{
See the \code{do_bind_rows} argument.
}
\description{
Convert list of lists to tibble.
}
\examples{
lists_to_tibble(list(list(a = 1, b = 2), list(b = 3, a = 4, c = 5)))
lists_to_tibble(list(list(a = 1, b = 2), list(b = 3, a = 4, c = 5)), do_bind_rows = FALSE)
}
\concept{misc_utils}
