% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans.R
\name{load_custom_plan}
\alias{load_custom_plan}
\title{Source a file returning a custom \link{drake} plan.}
\usage{
load_custom_plan(
  file = getOption("scdrake_plan_custom_file"),
  envir = parent.frame()
)
}
\arguments{
\item{file}{A character scalar: path to R script returning a \code{\link[drake:drake_plan]{drake::drake_plan()}}.}

\item{envir}{An environment in which the R script will be sourced (defaults to caller env).}
}
\value{
A \code{\link[drake:drake_plan]{drake::drake_plan()}} object (\code{tibble}).
}
\description{
This function is used in \link{drake}s init scripts \verb{_drake_single_sample.R} and \verb{_drake_integration.R}.
It sources an R script (\code{plan_custom.R} by default) which must return a \code{\link[drake:drake_plan]{drake::drake_plan()}} object.
In the R script, all variables defined in the init script are available, mainly \code{cfg} and \code{cfg_pipeline} lists
holding pipeline parameters. At the same time, all those variables are locked and cannot be modified in \code{file} script.
}
