% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_drake_functions.R
\name{LR_data_fn}
\alias{LR_data_fn}
\title{Receptor ligand analyses.}
\usage{
LR_data_fn(
  LR_data = LR_data,
  gobj = gobj,
  selected_clustering = selected_clustering
)
}
\arguments{
\item{LR_data}{Information about receptor and ligand pairng, inheritated from target from file contatining L-R pairing}

\item{gobj}{A \code{Giotto} object}

\item{selected_clustering}{A name of selected clustering for calculation}
}
\value{
\code{dataframe} containing probablities and other statistical information about L-R pairing
}
\description{
A calculation of ligand-receptor pairing probability
}
\concept{single_sample_norm_spatial}
