% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{make_gene_annotation}
\alias{make_gene_annotation}
\title{Create a \code{dataframe} with annotation of SCE object's genes using an \code{AnnotationDbi} package.}
\usage{
make_gene_annotation(sce, annotation_db_file)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object.}

\item{annotation_db_file}{A character scalar: path to SQLite file of annotation DB, e.g. \code{org.Hs.eg.db$conn@dbname}.}
}
\value{
A \code{dataframe} with annotation.
}
\description{
The following columns are retrieved from the \code{AnnotationDbi} package: \code{SYMBOL}, \code{GENENAME}, \code{ENTREZID}.
Multi-mapping entities are collapsed by "," (i.e. a single ENSEMBL ID having multiple symbols, etc.).
Unknown symbols are replaced by the corresponding ENSEMBL IDs.
}
\concept{sc_annotation}
