% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{make_kmeans_gaps_plot}
\alias{make_kmeans_gaps_plot}
\title{Make a plot of k-means gaps.}
\usage{
make_kmeans_gaps_plot(kmeans_gaps, best_k)
}
\arguments{
\item{kmeans_gaps}{An output from \code{\link[cluster:clusGap]{cluster::clusGap()}}.}

\item{best_k}{A numeric scalar: best k for k-means calculated by \code{\link[cluster:clusGap]{cluster::maxSE()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
For more details see \href{https://bioconductor.org/books/release/OSCA/clustering.html#base-implementation}{this}
chapter in OSCA.
}
\concept{sc_clustering}
