% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{make_pca_selected_pcs_plot}
\alias{make_pca_selected_pcs_plot}
\title{Make a plot showing numbers of selected PCs from all strategies.}
\usage{
make_pca_selected_pcs_plot(
  pca_percent_var,
  pca_elbow_pcs,
  pca_gene_var_pcs,
  pca_forced_pcs
)
}
\arguments{
\item{pca_percent_var}{The percentage of variance explained by each PC.
Obtained from \code{attr(reducedDim(sce_pca, "pca"), "percentVar")}.}

\item{pca_elbow_pcs}{A numeric scalar: number of PCs selected by elbow point strategy.}

\item{pca_gene_var_pcs}{A numeric scalar: number of PCs selected by technical noise strategy.}

\item{pca_forced_pcs}{A numeric scalar: constant number of PCs defined in config file.}
}
\value{
A \code{ggplot} object.
}
\description{
Make a plot showing numbers of selected PCs from all strategies.
}
\concept{sc_clustering}
