% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{make_sc3_stability_plots}
\alias{make_sc3_stability_plots}
\title{Create a list of plots with SC3 clustering stability.}
\usage{
make_sc3_stability_plots(sce_sc3, cluster_sc3, sc3_k)
}
\arguments{
\item{sce_sc3}{A \code{SingleCellExperiment} object with calculated SC3 clustering.}

\item{cluster_sc3}{A named list of SC3 cell-cluster assignments, as produced by \code{\link[=cluster_sc3_fn]{cluster_sc3_fn()}}.}

\item{sc3_k}{A numeric vector: numbers of clusters for SC3.}
}
\value{
A named list of \code{ggplot} objects. The names are the same as in the \code{cluster_sc3} function parameter.
}
\description{
For more details see \code{\link[SC3:sc3_plot_cluster_stability]{SC3::sc3_plot_cluster_stability()}}.

\code{metadata()} in \code{sce_sc3} are checked for \code{sc3_dry} item; if its value is \code{TRUE}, than SC3 clustering was run
in dry mode, producing random cell-cluster assignments, and thus, blank plots will be produced.
}
\concept{sc_clustering}
