% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_annotation.R
\name{makeSignMatrixPAGE}
\alias{makeSignMatrixPAGE}
\title{Create signature matrix from signature list.}
\usage{
makeSignMatrixPAGE(sign_names, sign_list)
}
\value{
Matrix: signature matrix based on markers in data
}
\description{
Create signature matrix from signature list.
}
\concept{manual_annotation}
