% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{marker_heatmap}
\alias{marker_heatmap}
\title{Make a heatmap of cell clusters.}
\usage{
marker_heatmap(
  seu,
  markers,
  groups,
  assay = "RNA",
  slot = "data",
  top_n = 5,
  top_n_wt = "fdr",
  title = "Heatmap",
  subtitle = "",
  fill_scale = ggplot2::scale_fill_gradient(low = "white", high = "red"),
  y_text_size = 8,
  angle = 0,
  ...
)
}
\arguments{
\item{seu}{A \code{Seurat} object.}

\item{markers}{A dataframe with markers (returned from \code{\link[=scran_markers]{scran_markers()}}) or a character vector of gene SYMBOLs.}

\item{groups}{A character vector: assignment of cells to groups (clusters).}

\item{assay, slot, angle, ...}{Passed to \code{\link[Seurat:DoHeatmap]{Seurat::DoHeatmap()}}.}

\item{top_n, top_n_wt}{Passed to \code{\link[=filter_markers]{filter_markers()}}. If \code{top_n} is \code{NULL}, all markers in \code{markers} will be used
(\code{SYMBOL} column in case \code{markers} is a dataframe-like object).
If \code{top_n} is numeric, \code{markers} must be a dataframe-like object.}

\item{title}{A character scalar: heatmap title.}

\item{subtitle}{A character scalar: heatmap subtitle.}

\item{fill_scale}{A \code{Scale} object from \code{ggplot2},
e.g. \code{ggplot2::scale_fill_gradient(low = "white", high = "red")} (default).}

\item{y_text_size}{A numeric scalar: text size on y-axe of the heatmap.}
}
\value{
A named list.
}
\description{
Make a heatmap of cell clusters.
}
\concept{sc_markers}
