% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{marker_heatmaps_wrapper}
\alias{marker_heatmaps_wrapper}
\title{Make a heatmap of global or contrast markers.}
\usage{
marker_heatmaps_wrapper(
  seu,
  params,
  marker_type = c("global", "contrast"),
  save = TRUE,
  return_type = c("file", "tibble")
)
}
\arguments{
\item{seu}{A \code{Seurat} object.}

\item{params}{A tibble. See \code{cluster_markers_heatmap_params} or \code{contrasts_heatmap_params} target for details.}

\item{marker_type}{A character scalar.}

\item{save}{A logical scalar. If \code{TRUE}, the heatmap will be saved to \code{params$out_file}.}

\item{return_type}{A character scalar. For \code{"file"}, return \code{params$out_file}.
For \code{"tibble"}, return a modified \code{params} tibble with appended list columns:
\itemize{
\item \code{heatmaps}: contains named lists of heatmap objects (\code{p_heatmap}, \code{p_heatmap_z_score})
\item \code{markers_top}: dataframe with top markers used for heatmap generation.
}}
}
\value{
See the \code{return_type} argument.
}
\description{
Make a heatmap of global or contrast markers.
}
\concept{sc_markers}
