% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{marker_plot}
\alias{marker_plot}
\title{Make a marker plot.}
\usage{
marker_plot(
  sce,
  dimred_name,
  gene_ensembl_id,
  group,
  cluster_plot_title = NULL,
  cluster_plot_subtitle = NULL,
  cluster_plot_legend_title = NULL,
  cluster_plot_label = TRUE,
  cluster_plot_point_size = 1,
  feature_plot_point_size = 1,
  vln_plot_legend_title = NULL
)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object.}

\item{dimred_name}{A character scalar.}

\item{gene_ensembl_id}{A character scalar. ENSEMBL ID of a marker.}

\item{group}{\itemize{
\item A character scalar: name of column in \code{colData(sce)} by which are cells separated (e.g. clustering).
\item A character vector or factor: assignment of cells to groups (clusters). Must have the same length as the number
of cells (columns) in \code{sce}.
}}

\item{cluster_plot_title, cluster_plot_subtitle, cluster_plot_legend_title, cluster_plot_label}{A character scalar. Applied to the dimred plot which is colored by \code{group}.}

\item{cluster_plot_point_size}{A numeric scalar.}

\item{feature_plot_point_size}{A numeric scalar. Applied to the feature plot.}

\item{vln_plot_legend_title}{A character scalar. Applied to the violin plot.}
}
\value{
A \code{patchwork} object.
}
\description{
A marker plot is composed of:
\itemize{
\item Two plots of a dimred: one colored by a clustering, and second colored by expression of a marker (feature plot).
\item A plot of summarized marker expression: proportion of cells expressing a marker, colored by average expression.
\item A violin plot: expression of a marker in each cluster.
}
}
\concept{sc_markers}
