% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{markers_dimred_plots}
\alias{markers_dimred_plots}
\title{Make dimred plots for each marker source defined in \code{CLUSTER_MARKERS_SOURCES} (\code{cluster_markers.yaml} config)
or \code{CONTRASTS_SOURCES} (\code{contrasts.yaml} config).}
\usage{
markers_dimred_plots(
  sce_final_norm_clustering,
  markers_dimred_plots_params,
  spatial = FALSE
)
}
\arguments{
\item{sce_final_norm_clustering}{A \code{SingleCellExperiment} object with computed dimreds and clusterings.}

\item{markers_dimred_plots_params}{A tibble. See \code{cluster_markers_dimred_plot_params} or
\code{contrasts_dimred_plot_params} for details.}
}
\value{
A modified \code{markers_dimred_plots_params} tibble with appended \code{plot} column (list of \code{ggplot2} objects).
See \code{cluster_markers_dimred_plots} or \code{contrasts_dimred_plots} target for details.
}
\description{
Make dimred plots for each marker source defined in \code{CLUSTER_MARKERS_SOURCES} (\code{cluster_markers.yaml} config)
or \code{CONTRASTS_SOURCES} (\code{contrasts.yaml} config).
}
\concept{sc_markers}
