% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{markers_dimred_plots_files}
\alias{markers_dimred_plots_files}
\title{Save dimred plots.}
\usage{
markers_dimred_plots_files(
  markers_dimred_plots,
  markers_dimred_plots_base_out_dir
)
}
\arguments{
\item{markers_dimred_plots}{A tibble. See \code{cluster_markers_dimred_plots} or \code{contrasts_dimred_plots} for details.}

\item{markers_dimred_plots_base_out_dir}{A character scalar: path to base output directory to save dimred plots in.}
}
\value{
A modified \code{markers_dimred_plots} tibble with appended \code{out_file} column.
See \code{cluster_markers_dimred_plots_files} or \code{contrasts_dimred_plots_files} target for details.
}
\description{
Save dimred plots.
}
\concept{sc_markers}
