% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{markers_for_tables}
\alias{markers_for_tables}
\title{Make a table used in HTML report of marker results.}
\usage{
markers_for_tables(
  markers_out,
  markers_heatmaps_df,
  markers_plots_top,
  markers_dimred_plots_files,
  out_dir,
  ensembl_species = "Homo_sapiens",
  ...
)
}
\arguments{
\item{markers_out}{A tibble. See \code{cluster_markers_out} or \code{contrasts_out} target for details.}

\item{markers_heatmaps_df}{A tibble. See \code{cluster_markers_heatmaps_df} or \code{contrasts_heatmaps_df} for details.}

\item{markers_plots_top}{A tibble.
See \code{cluster_markers_plots_top} or \code{contrasts_plots_top} for details.}

\item{markers_dimred_plots_files}{A tibble.
See \code{cluster_markers_dimred_plots_files} or \code{contrasts_dimred_plots_files} for details.}

\item{out_dir}{A character scalar: path to base output directory to save HTML tables in.
Files and subdirectories will be created as:
\itemize{
\item \verb{<source_name>/<source_name>_<test_type>_<group_level>.html} for cluster markers.
Example: \code{markers_cluster_kmeans_kc_4/markers_cluster_kmeans_kc_4_t_1.html}
\item \verb{<source_name>/<source_name>_<test_type>_<target_level>_vs_<reference_level>.html} for contrasts.
Example: \code{dea_cluster_louvain_annotated/dea_cluster_louvain_annotated_binom_cl3_vs_cl4.html}
}}

\item{ensembl_species}{A character scalar: passed to \code{\link[=add_marker_table_links]{add_marker_table_links()}}.}

\item{...}{Currently not used.}
}
\value{
A tibble. See \code{cluster_markers_for_tables} or \code{contrasts_for_tables} for details.
}
\description{
It combines marker statistics, parameters, dimred plots, and links to heatmaps.
}
\concept{sc_markers}
