% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{markers_plots_files}
\alias{markers_plots_files}
\title{Make plots of top markers.}
\usage{
markers_plots_files(
  sce_dimred,
  markers_plots_top,
  save = TRUE,
  dry = FALSE,
  spatial = FALSE,
  return_type = c("file", "plot", "tibble")
)
}
\arguments{
\item{sce_dimred}{A \code{SingleCellExperiment} object with computed dimreds.}

\item{markers_plots_top}{A tibble, output from \code{\link[=markers_plots_top]{markers_plots_top()}}.
See \code{cluster_markers_plots_top} or \code{contrasts_plots_top} target for details.}

\item{save}{A logical scalar: if \code{TRUE}, save each plot to \code{markers_top$out_file}.}

\item{dry}{A logical scalar: if \code{TRUE}, do not make plots and just create empty files.}

\item{return_type}{A character scalar:
\itemize{
\item \code{"file"}: return \code{markers_top$out_file}.
\item \code{"plot"}: return list of marker plots.
\item \code{"tibble"}: return a modified \code{markers_plots_top} tibble with appended \code{marker_plot} column (list of marker plots).
}}
}
\value{
See the \code{return_type} argument.
See \code{cluster_markers_plots_files} or \code{contrasts_plots_files} target for details.
}
\description{
Make plots of top markers.
}
\seealso{
\code{\link[=marker_plot]{marker_plot()}}
}
\concept{sc_markers}
