% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{markers_plots_top}
\alias{markers_plots_top}
\title{Get a tibble of top markers extracted from each test result.}
\usage{
markers_plots_top(
  markers_processed,
  markers_plot_params,
  out_dir,
  marker_type = c("global", "contrast")
)
}
\arguments{
\item{markers_processed}{A tibble. See \code{cluster_markers_processed} or \code{contrasts} target for details.}

\item{markers_plot_params}{A tibble. See \code{cluster_markers_plot_params} or \code{contrasts_plot_params} for details.}

\item{out_dir}{A character scalar: path to base output directory for plots.
Subdirectories and plot files will be created as
\verb{<clustering_name>/<dimred>/<ENSEMBL>_<SYMBOL>_<clustering_name>_<dimred>.pdf}.}

\item{marker_type}{A character scalar.}
}
\value{
A tibble. See \code{cluster_markers_plots_top} or \code{contrasts_plots_top} target for details.
}
\description{
This is basically an unnesting of filtered \code{markers} in each row of \code{markers_processed}.
The filtering is based on \code{top_n_plot} and \code{top_n_wt_plot} columns in \code{markers_plot_params}.
}
\concept{sc_markers}
