% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{markers_table_files}
\alias{markers_table_files}
\title{Prepare a table with markers and render a HTML report from RMarkdown template.}
\usage{
markers_table_files(
  markers_for_tables,
  rmd_template,
  marker_type = c("global", "contrast"),
  drake_cache_dir = ".drake",
  spatial = FALSE
)
}
\arguments{
\item{markers_for_tables}{A tibble. File paths in column \code{out_file} will be used to save rendered HTML files.
See \code{cluster_markers_for_tables} or \code{contrasts_for_tables} target for details.}

\item{rmd_template}{A character scalar: path to RMarkdown file.}

\item{marker_type}{A character scalar.}

\item{drake_cache_dir}{A character scalar: path to \code{drake} cache directory.}
}
\value{
A character vector: \code{markers_for_tables$out_file}.
}
\description{
Prepare a table with markers and render a HTML report from RMarkdown template.
}
\concept{sc_markers}
