% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{md_header}
\alias{md_header}
\title{Generate a Markdown header.}
\usage{
md_header(text, heading, extra = "", do_cat = TRUE)
}
\arguments{
\item{text}{A character scalar: text of the header.}

\item{heading}{A numeric scalar: level of the header, that is, the result will be \code{"#"} times \code{heading}.}

\item{extra}{A character scalar: content to put after \code{text}, e.g. \code{"{.tabset}"} for tabbed heading.}

\item{do_cat}{A logical scalar: if \code{TRUE}, print (using \code{cat()}) the result before returning.}
}
\value{
A character scalar.
}
\description{
Generate a Markdown header.
}
\concept{misc_html}
