% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce.R
\name{merge_sce_metadata}
\alias{merge_sce_metadata}
\title{Merge \code{metadata()} of multiple \code{SingleCellExperiment} object.}
\usage{
merge_sce_metadata(sce_list, what, as_vector = TRUE, as_named_list = FALSE)
}
\arguments{
\item{sce_list}{A list of \code{SingleCellExperiment} objects.}

\item{what}{A character scalar: name of item to merge.}

\item{as_vector}{A logical scalar: if \code{TRUE}, return vector of merged values.}

\item{as_named_list}{A logical scalar: if \code{TRUE}, return a named list with name equal to \code{what}.}
}
\value{
A vector or list.
}
\description{
Merge \code{metadata()} of multiple \code{SingleCellExperiment} object.
}
\concept{sc_sce}
