% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_annotation.R
\name{meta_heatmap_ploting}
\alias{meta_heatmap_ploting}
\title{Manual annotation heatmap plotting}
\usage{
meta_heatmap_ploting(
  sce,
  clus_cor_method = "pearson",
  clus_cluster_method = "complete",
  values_cor_method = "pearson",
  values_cluster_method = "complete",
  clustering,
  gradient_midpoint = 0,
  gradient_limits = NULL,
  x_text_size = 10,
  x_text_angle = 45,
  y_text_size = 10,
  strip_text_size = 8,
  low = "blue",
  mid = "white",
  high = "red",
  spatial = FALSE,
  make_cell_plot = FALSE,
  out_dir = NULL
)
}
\arguments{
\item{sce}{A \code{SingleCellAnnotation} object}

\item{clustering}{Selected clustering}

\item{spatial}{Logical vector, if include spot images for each anotation}

\item{make_cell_plot}{Logical vector, if include pseudotissue images, for spatial extension}
}
\description{
Manual annotation heatmap plotting
}
\concept{manual_annotation}
