% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{na_empty}
\alias{na_empty}
\title{Return \code{NA} if an object is empty.}
\usage{
na_empty(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{NA} if \code{x} is empty (\code{length(x) == 0}), otherwise \code{x}.
}
\description{
Return \code{NA} if an object is empty.
}
\examples{
na_empty(integer())
}
\concept{misc_utils}
