% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_sample_norm_clustering.R
\name{pca_phase_plots_fn}
\alias{pca_phase_plots_fn}
\title{Make dimred plots of PCA colored by cell cycle phase.}
\usage{
pca_phase_plots_fn(sce_pca)
}
\arguments{
\item{sce_pca}{(\emph{input target}) A \code{SingleCellExperiment} object.}
}
\value{
A list of one or two \code{ggplot2} objects. \emph{Output target}: \code{pca_phase_plots}
}
\description{
If HVG selection was performed with removal of CC-related genes, list of two plots will be returned -
one with PCA performed on all genes, and the second one without CC-related genes.
}
\concept{single_sample_norm_clustering_fn}
