% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_clustree}
\alias{plot_clustree}
\title{Plot clustering tree.}
\usage{
plot_clustree(
  cluster_list,
  params,
  prefix,
  title = deparse(substitute(cluster_list)),
  edge_arrow = FALSE,
  highlight_core = TRUE,
  ...
)
}
\arguments{
\item{cluster_list}{A named list of vectors.}

\item{params}{A vector of clustering parameters (resolutions, \code{k}s) in the same order as \code{cluster_list}.}

\item{prefix, edge_arrow, highlight_core, ...}{Passed to \code{\link[clustree:clustree]{clustree::clustree()}}.}

\item{title}{A character scalar: plot title.}
}
\value{
A \code{ggplot} object.
}
\description{
\code{\link[clustree:clustree]{clustree::clustree()}} is used under the hood.
}
\concept{sce_visualization}
