% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_colData}
\alias{plot_colData}
\title{A wrapper for \code{\link[scater:plotColData]{scater::plotColData()}}.}
\usage{
plot_colData(
  ...,
  title = "",
  scale_x = NULL,
  scale_y = NULL,
  add_logical_counts = TRUE
)
}
\arguments{
\item{...}{Parameters passed to \code{\link[scater:plotColData]{scater::plotColData()}}.}

\item{title}{A characater scalr: main title of plot.}

\item{scale_x, scale_y}{A \code{ggplot2} scale, e.g. \code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_log10()}}.}

\item{add_logical_counts}{A logical scalar: if \code{TRUE}, add number of each logical level
(\code{TRUE}, \code{FALSE}) used for coloring.}
}
\value{
A \code{ggplot2} object.
}
\description{
Can add title, scales and number of each logical level (\code{TRUE}, \code{FALSE}) used for coloring.
}
\concept{sce_visualization}
