% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{plot_deconv_results_fn}
\alias{plot_deconv_results_fn}
\title{Deconvolution annotation plotting.}
\usage{
plot_deconv_results_fn(spe, outdir)
}
\arguments{
\item{spe}{A \code{SingleCellAnnotation} object.}
}
\value{
List of images.
}
\description{
Deconvolution annotation plotting.
}
\concept{deconvolution}
