% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hvg.R
\name{plot_hvg}
\alias{plot_hvg}
\title{Plot average expression vs. HVG metrics (total, bio, and technical variance) and highlight HVGs.}
\usage{
plot_hvg(sce, ...)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object with HVG metadata.}

\item{...}{Currently not used.}
}
\value{
A \code{ggplot2} object if \code{hvg_metric} is \code{sctransform}, otherwise a \code{patchwork} object.
}
\description{
Plot average expression vs. HVG metrics (total, bio, and technical variance) and highlight HVGs.
}
\concept{sc_hvg}
