% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hvg.R
\name{plot_hvg_fit}
\alias{plot_hvg_fit}
\title{Plot fit of a HVG metric and highlight HVGs.}
\usage{
plot_hvg_fit(sce, y = c("var", "cv2"))
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object with HVG metadata.}

\item{y}{A character scalar: which column from HVG fit \code{DataFrame} to plot on the y-axe.
\code{var} or \code{cv2} if \code{hvg_metric} is \code{gene_var} or \code{gene_cv2}, respectively.}
}
\value{
Invisibly \code{NULL} (drawing on a graphical device).
}
\description{
Plot fit of a HVG metric and highlight HVGs.
}
\concept{sc_hvg}
