% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visualization.R
\name{plot_spat_visuals}
\alias{plot_spat_visuals}
\title{A function for visualization selected qc matrices in pseudotissue visualization}
\usage{
plot_spat_visuals(sce)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object.}
}
\value{
A list of plots.
}
\description{
Adapted function from Giotto package \link{Dries et al, 2021}, rewrite for use in scdrake package for a SingleCellExperiment object. Helper function for users, not in core scdrake package.
}
\concept{spatial_visualization}
