% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visualization.R
\name{plot_spat_visuals2}
\alias{plot_spat_visuals2}
\title{A function for visualization selected qc matrices in pseudotissue}
\usage{
plot_spat_visuals2(spe)
}
\arguments{
\item{sce}{A \code{SpatialExperiment} object.}
}
\value{
A list of plots.
}
\description{
Plot spots features as spots
}
\concept{spatial_visualization}
