% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_vln}
\alias{plot_vln}
\title{Make a combined violin-boxplot plot.}
\usage{
plot_vln(
  sce,
  ensembl_id,
  groups,
  add_jitter = TRUE,
  jitter_color = TRUE,
  jitter_point_size = 0.5,
  violin_scale = "area",
  boxplot_width = 0.1,
  labs = NULL
)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object.}

\item{ensembl_id}{A character scalar: ENSEMBL ID of feature to plot.}

\item{groups}{A character scalar: name of a column in \code{colData(sce)} by which to divide cells.
Violin plot will be created for each level.}

\item{add_jitter}{A logical scalar: if \code{TRUE}, add jitter to points.}

\item{jitter_color}{A logical scalar: if \code{TRUE}, jittered points will be colored by \code{groups}.}

\item{jitter_point_size}{A numeric scalar.}

\item{violin_scale}{A character scalar: passed to \code{\link[ggplot2:geom_violin]{ggplot2::geom_violin()}} (\code{scale} parameter).}

\item{boxplot_width}{A numeric scalar: passed to \code{\link[ggplot2:geom_boxplot]{ggplot2::geom_boxplot()}} (\code{width} parameter).}

\item{labs}{A \code{\link[ggplot2:labs]{ggplot2::labs()}} object.}
}
\value{
A \code{ggplot2} object.
}
\description{
Make a combined violin-boxplot plot.
}
\concept{sce_visualization}
