% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_drake_functions.R
\name{plotCCcomHeatmap_fn}
\alias{plotCCcomHeatmap_fn}
\title{Heatmap visulization of selected L-R pairing}
\usage{
plotCCcomHeatmap_fn(
  gobject = gobject,
  comScores = comScores,
  selected_LR = selected_LR,
  selected_cell_LR = selected_cell_LR,
  out_dir = out_dir
)
}
\arguments{
\item{gobject}{A \code{Giotto} object}

\item{comScores}{Precalculated L-R pairing scores from LR_data_fn}

\item{selected_LR}{Precalculated individual expression scores from top_LR_ints_fn_new}

\item{selected_cell_LR}{Precalculated combined cell scores from top_LR_cell_ints_fn}

\item{out_dir}{Name of output directory in fs file system from here pathway}
}
\description{
Heatmap visualization of selected most significant L-R pairing, based on PI values
}
\concept{single_sample_norm_spatial}
