% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotReducedDim_mod}
\alias{plotReducedDim_mod}
\title{A wrapper around \code{\link[scater:plotReducedDim]{scater::plotReducedDim()}}.}
\usage{
plotReducedDim_mod(
  sce,
  dimred,
  colour_by = NULL,
  point_size = 1,
  point_alpha = 0.8,
  title = "",
  subtitle = "",
  use_default_ggplot_palette = TRUE,
  colour_gradient = list(low = "lightgrey", high = "blue"),
  legend_title = NULL,
  add_cells_per_cluster = TRUE,
  ...
)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object.}

\item{dimred, colour_by, point_size, point_alpha, ...}{Passed to \code{\link[scater:plotReducedDim]{scater::plotReducedDim()}}.}

\item{title, subtitle, legend_title}{A character scalar.}

\item{use_default_ggplot_palette}{A logical scalar: if \code{TRUE}, use the default \code{ggplot2} palette
(\code{\link[ggplot2:scale_hue]{ggplot2::scale_color_hue()}}).}

\item{colour_gradient}{A named list with low and high colors for continuous variables.}

\item{add_cells_per_cluster}{A logical scalar: if \code{TRUE} and \code{colour_by} variable is (coercible to) factor,
add number of cells per each level to color legend labels.}
}
\value{
A \code{ggplot2} object.
}
\description{
A wrapper around \code{\link[scater:plotReducedDim]{scater::plotReducedDim()}}.
}
\concept{sce_visualization}
