% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_process_common.R
\name{.prepare_marker_source_params}
\alias{.prepare_marker_source_params}
\alias{.prepare_marker_sources_params}
\title{Prepare a \code{tibble} with parameters for cluster marker tests.}
\usage{
.prepare_marker_source_params(
  marker_source,
  marker_source_name,
  marker_default_params
)

.prepare_marker_sources_params(marker_sources, marker_default_params)
}
\arguments{
\item{marker_source}{A named list.}

\item{marker_source_name}{A character scalar: name of a marker source.}

\item{marker_default_params}{A named list with default parameters (common and for tests).
Obtained from \code{CLUSTER_MARKERS_SOURCES_DEFAULTS} or \code{CONTRASTS_SOURCES_DEFAULTS}.}

\item{marker_sources}{A named list of named lists of marker sources.}
}
\value{
A named list (of named lists).
}
\description{
Prepare a \code{tibble} with parameters for cluster marker tests.
}
\concept{internal}
