% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{prepare_markers_source_params}
\alias{prepare_markers_source_params}
\alias{prepare_markers_sources_params}
\title{Prepare a list of marker source parameters.}
\usage{
prepare_markers_source_params(
  marker_source,
  marker_source_name,
  marker_default_params
)

prepare_markers_sources_params(marker_sources, marker_default_params)
}
\arguments{
\item{marker_source}{A named list.}

\item{marker_source_name}{A character scalar.}

\item{marker_default_params}{A named list.}

\item{marker_sources}{A list of marker sources.}
}
\value{
A modified \code{marker_source} list.

A modified \code{marker_sources} list.
}
\description{
Prepare a list of marker source parameters.
}
\concept{sc_markers}
