% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_process_common.R,
%   R/config_process_integration.R, R/config_process_single_sample.R
\name{process_config}
\alias{process_config}
\alias{.process_pipeline_config}
\alias{.process_main_config}
\alias{.process_cluster_markers_config}
\alias{.process_contrasts_config}
\alias{.prepare_cell_annotation_sources_params}
\alias{.process_integration_config}
\alias{.process_int_clustering_config}
\alias{.process_input_qc_config}
\alias{.process_norm_clustering_config}
\title{Config processing.}
\usage{
.process_pipeline_config(cfg)

.process_main_config(cfg)

.process_cluster_markers_config(cfg, other_variables)

.process_contrasts_config(cfg, other_variables)

.prepare_cell_annotation_sources_params(
  cell_annotation_sources,
  cell_annotation_default_params
)

.process_integration_config(cfg, other_variables)

.process_int_clustering_config(cfg, other_variables)

.process_input_qc_config(cfg, other_variables)

.process_norm_clustering_config(cfg, other_variables)
}
\arguments{
\item{cfg}{A named list with pipeline parameters.}

\item{other_variables}{A list of variables in whose environment the \code{yaml_file} will be evaluated,
see details in \code{\link[=load_config]{load_config()}}.}

\item{cell_annotation_sources}{A named list.}

\item{cell_annotation_default_params}{A named list with default parameters.}
}
\value{
A modified \code{cfg}.
}
\description{
These functions are responsible for checking and processing of config parameters. The most important steps are:
\itemize{
\item Setting up the pipeline parameters, such as \code{drake} targets and cache, parallel processing, etc.
\item Extracting a path to SQLite database file from gene annotation package.
\item Conversion of relative paths (to project's root directory) to absolute and putting some paths under a base output directory.
\item Checking and processing of \code{CLUSTER_MARKERS_SOURCES}, \code{CONTRASTS_SOURCES}, and \code{INTEGRATION_SOURCES} parameters.
}
}
\concept{internal}
