% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{render_bootstrap_table}
\alias{render_bootstrap_table}
\title{Render a dataframe-like object using \code{\link[knitr:kable]{knitr::kable()}} and \code{\link[kableExtra:kable_styling]{kableExtra::kable_styling()}}.}
\usage{
render_bootstrap_table(
  df,
  bootstrap_options = c("striped", "hover", "condensed"),
  full_width = TRUE,
  position = "center",
  row.names = TRUE,
  ...
)
}
\arguments{
\item{df}{A dataframe-like object.}

\item{bootstrap_options, full_width, position}{Passed to \code{\link[kableExtra:kable_styling]{kableExtra::kable_styling()}}.}

\item{row.names}{Passed to \code{\link[knitr:kable]{knitr::kable()}}.}

\item{...}{Passed to \code{\link[knitr:kable]{knitr::kable()}} and \code{\link[kableExtra:kable_styling]{kableExtra::kable_styling()}}.}
}
\value{
An object of class \code{kableExtra} and \code{knitr_kable}.
}
\description{
Render a dataframe-like object using \code{\link[knitr:kable]{knitr::kable()}} and \code{\link[kableExtra:kable_styling]{kableExtra::kable_styling()}}.
}
\concept{misc_html}
