% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_list_nas_with_nulls}
\alias{replace_list_nas_with_nulls}
\title{Replace \code{NA}s in a list by \code{NULL}s.}
\usage{
replace_list_nas_with_nulls(l)
}
\arguments{
\item{l}{A list.}
}
\value{
The list \code{l} with \code{NA}s replaced by \code{NULL}s.
}
\description{
Replace \code{NA}s in a list by \code{NULL}s.
}
\examples{
replace_list_nas_with_nulls(list(NA, a = 1, b = NA))

}
\concept{misc_utils}
