% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_list_nulls}
\alias{replace_list_nulls}
\title{Replace \code{NULL} items in a defined depth of a list.}
\usage{
replace_list_nulls(l, replacement = NA, depth = 2L)
}
\arguments{
\item{l}{A list.}

\item{replacement}{A replacement for \code{NULL} items.}

\item{depth}{An integer scalar: depth of \code{l} to search for and replace \code{NULL} items.}
}
\value{
The list \code{l} with \code{NULL} items replaced by \code{replacement} in the defined \code{depth}.
}
\description{
Replace \code{NULL} items in a defined depth of a list.
}
\examples{
replace_list_nulls(list(a = list(b = 1, c = NULL)))
}
\concept{misc_utils}
