% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{run_kmeans_clustering}
\alias{run_kmeans_clustering}
\title{Run k-means clustering for a specific \code{k}.}
\usage{
run_kmeans_clustering(
  sce,
  kmeans_k,
  is_integration,
  dimred = "pca",
  nstart = 25,
  iter.max = 1000
)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object.}

\item{kmeans_k}{An integer scalar: number of clusters for k-means.}

\item{is_integration}{A logical scalar: if \code{TRUE}, clusterings will be named as \verb{cluster_int_*}, otherwise \verb{cluster_*}.}

\item{dimred}{A character scalar: name of matrix in \code{reducedDim()} used for k-means.}

\item{nstart, iter.max}{Passed to \code{\link[stats:kmeans]{stats::kmeans()}}.}
}
\value{
A \code{tibble} whose columns are mostly self-explanatory, except the \code{data} column, which is of \code{list} type and
contains an another \code{tibble} with \code{kmeans_object} column holding an object of class \code{kmeans} returned from
the \code{\link[stats:kmeans]{stats::kmeans()}} function.
}
\description{
Run k-means clustering for a specific \code{k}.
}
\concept{sc_clustering}
