% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_annotation.R
\name{run_page_man_annotation}
\alias{run_page_man_annotation}
\title{Calculate and run PAGE annotation.}
\usage{
run_page_man_annotation(
  sign_matrix,
  sce,
  values = "logcounts",
  scale = NULL,
  overlap = 5,
  reverse_log_scale = FALSE,
  selected_annotation = NULL,
  output_enrichment = "zscore"
)
}
\arguments{
\item{sign_matrix}{precalculated signature matrix}

\item{sce}{A \code{SingleCellExperiment} object}

\item{values}{A expresion indicating which values use, logcounts as default}
}
\description{
Calculate and run PAGE annotation.
}
\concept{manual_annotation}
