% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{save_clustree}
\alias{save_clustree}
\title{Save a clustree plot into PDF.}
\usage{
save_clustree(p, out_file, width = 14, height = 10, ...)
}
\arguments{
\item{p}{A \code{ggplot} object.}

\item{out_file}{A character scalar: output PDF file.}

\item{width, height, ...}{Passed to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}.}
}
\value{
A character scalar: \code{out_file}
}
\description{
Save a clustree plot into PDF.
}
\concept{sce_visualization}
