% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_pdf}
\alias{save_pdf}
\title{Save a list of plots to multipage PDF.}
\usage{
save_pdf(
  plots,
  output_file,
  width = NULL,
  height = NULL,
  make_thumbnail = FALSE
)
}
\arguments{
\item{plots}{A list of plots.}

\item{output_file}{A character scalar: path to output PDF file.
If file's directory doesn't exist, it will be created recursively.}

\item{width, height}{A numeric scalar: default width and height of graphics region in inches. Defaults to 7.}

\item{make_thumbnail}{A logical scalar: if \code{TRUE}, a PNG file will be created from the first plot in \code{plots}.}
}
\value{
A character scalar (\code{output_file}) or vector of length two (\code{c(output_file, thumbnail_file)}) if
\code{make_thumbnail} is \code{TRUE}.
}
\description{
Save a list of plots to multipage PDF.
}
\concept{misc_utils}
