% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_print}
\alias{save_print}
\alias{save_object_info}
\title{Capture output from \code{print()} of an object.}
\usage{
save_print(object, collapse = "\\n")

save_object_info(object)
}
\arguments{
\item{object}{An object to be printed.}

\item{collapse}{A character scalar: used to collapse the output lines.}
}
\value{
A character scalar for \code{save_print()},
\code{list(str = save_print(object), dim = dim(object))} for \code{save_object_info()}.
}
\description{
Capture output from \code{print()} of an object.
}
\examples{
save_print(lm(mpg ~ cyl, datasets::mtcars))
}
\concept{misc_utils}
