% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdrake-package.R
\docType{data}
\name{SCDRAKE_CLI_VERSION}
\alias{SCDRAKE_CLI_VERSION}
\title{Matching CLI version for this package version.}
\format{
An object of class \code{character} of length 1.
}
\usage{
SCDRAKE_CLI_VERSION
}
\description{
Matching CLI version for this package version.
}
\concept{cli}
\keyword{datasets}
