% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_scdrake_default_options}
\alias{get_scdrake_default_options}
\alias{get_scdrake_options}
\title{Options used by \code{scdrake}.}
\usage{
get_scdrake_default_options(verbose = TRUE)

get_scdrake_options()
}
\arguments{
\item{verbose}{A logical scalar: if \code{TRUE}, be verbose.}
}
\value{
A named list of options.
}
\description{
\code{get_scdrake_default_options()} returns a list of default \code{scdrake} options passed to
\code{options()} during the package load.

\code{get_scdrake_options()} returns a list of options currently used by \code{scdrake}
(their names begin with \code{"scdrake_"}).
}
\details{
Most of the options are obtained from environment variables named in UPPERCASE,
e.g. the value of \code{scdrake_verbose} is taken from \code{SCDRAKE_VERBOSE}.
The environment variables are coerced to atomic types (character, logical, integer, double).

The following options are used by \code{scdrake}:
\itemize{
\item \code{scdrake_yq_binary}
(character, default: \code{Sys.which("yq")})
\itemize{
\item A path to \code{yq} tool's binary.
\item If the default value is \code{""}, it will be set to \code{get_yq_default_path()}.
}
\item \code{scdrake_verbose}
(logical, env: \code{SCDRAKE_VERBOSE}, default: \code{TRUE})
\itemize{
\item If \code{TRUE}, \code{scdrake} will be verbose.
}
\item \code{scdrake_cache_dir}
(character, env: \code{SCDRAKE_CACHE_DIR}, default: \code{".drake"})
\itemize{
\item A path to \code{drake} cache directory.
The \code{drake} package also contains an RStudio addin, which has a shortcut for loading a target under cursor
(see \verb{Tools -> Modify Keyboard Shortcuts -> "loadd target under cursor"}).
To know which cache directory to use for loading, \code{drake} is using the \code{rstudio_drake_cache} option.
On its load or attach, \code{scdrake} will automatically set this option to \code{scdrake_cache_dir}.
Just keep in mind that \code{rstudio_drake_cache} option will be overwritten if you load the \code{drake} package.
}
\item \code{scdrake_pipeline_config_dir}
(character, env: \code{SCDRAKE_PIPELINE_CONFIG_DIR}, default: \code{"config"})
\itemize{
\item A path to directory containing \code{pipeline.default.yaml} and/or \code{pipeline.yaml} files.
}
\item \code{scdrake_single_sample_config_dir}
(character, env: \code{SCDRAKE_SINGLE_SAMPLE_CONFIG_DIR}, default: \code{"config/single_sample"})
\itemize{
\item A path to directory containing files for single-sample analysis,
see the \code{dir} parameter in \code{\link[=load_single_sample_configs]{load_single_sample_configs()}}.
}
\item \code{scdrake_integration_config_dir}
(character, env: \code{SCDRAKE_INTEGRATION_CONFIG_DIR}, default: \code{"config/integration"})
\itemize{
\item A path to directory containing files for integration analysis,
see the \code{dir} parameter in \code{\link[=load_integration_configs]{load_integration_configs()}}.
}
\item \code{scdrake_plan_custom_file}
(character, env: \code{SCDRAKE_PLAN_CUSTOM_FILE}, default: \code{"plan_custom.R"})
\itemize{
\item A path to file with custom \code{drake} plan.
}
\item \code{scdrake_project_root}
(character, env: \code{SCDRAKE_PROJECT_ROOT}, default: \code{.})
\itemize{
\item A path to \code{scdrake} project root. If different from the current working directory (\code{.}), \code{scdrake} will first
change the working directory before pipeline is run in e.g. \code{run_single_sample_r()} or \code{run_single_sample()}.
That means other path-specifying options (e.g. \code{scdrake_pipeline_config_dir}) and config parameters will become
relative to \code{scdrake_project_root} - \strong{use at your own risk}!
}
}
}
\concept{scdrake_options}
