% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce.R
\name{sce_add_cell_data}
\alias{sce_add_cell_data}
\title{Add columns to \code{colData()} of a \code{SingleCellExperiment} object.}
\usage{
sce_add_cell_data(sce_dimred, cell_data)
}
\arguments{
\item{sce_dimred}{A \code{SingleCellExperiment} object.}

\item{cell_data}{A dataframe.}
}
\value{
A modified \code{sce_dimred} with added \code{colData()} columns from \code{cell_data}.
}
\description{
Only columns unique to \code{cell_data} will be added.
}
\concept{sc_sce}
