% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce.R
\name{sce_add_metadata}
\alias{sce_add_metadata}
\title{Append data to \code{metadata()} list of a \code{SingleCellExperiment} object.}
\usage{
sce_add_metadata(sce, ...)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object.}

\item{...}{Objects to be added to \code{metadata(sce)}.}
}
\value{
A modified \code{sce} object with data appended to its \code{metadata()}.
}
\description{
\code{\link[utils:modifyList]{utils::modifyList()}} is used internally, so existing named items in \code{metadata()} can be overwritten.
}
\examples{
\dontrun{
sce <- sce_add_metadata(sce, a = 1, "some_data", b = list(c = 1))
}

}
\concept{sc_sce}
