% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce.R
\name{sce_add_spatial_colData}
\alias{sce_add_spatial_colData}
\title{Append new columns with spatial relevance to \code{colData} of a \code{SingleCellExperiment} object.}
\usage{
sce_add_spatial_colData(sce, spatial_locs, spatial = FALSE)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object.}

\item{spatial_locs}{A file contating spatial coordiantes}

\item{spatial}{Logical vector If true, add spatial coordinates}
}
\description{
Append new columns with spatial relevance to \code{colData} of a \code{SingleCellExperiment} object.
}
\concept{spatial_sce}
