% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{sce_calc_pca}
\alias{sce_calc_pca}
\title{Calculate PCA of a \code{SingleCellExperiment} object.}
\usage{
sce_calc_pca(
  sce,
  name = "pca",
  exprs_values = "logcounts",
  subset_row = NULL,
  BSPARAM = BiocSingular::IrlbaParam(),
  BPPARAM = BiocParallel::SerialParam(),
  ...
)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object.}

\item{name, exprs_values, BSPARAM, BPPARAM, ...}{Passed to \code{\link[scater:runPCA]{scater::runPCA()}}.}

\item{subset_row}{A character vector. If \code{NULL}, IDs of highly variable genes stored in \code{metadata(sce)$hvg_ids} will be used for
the \code{subset_row} parameter in \code{\link[scater:runPCA]{scater::runPCA()}}. If HVGs are not found, all genes will be used.}
}
\value{
A \code{SingleCellExperiment} object with calculated PCA.
}
\description{
Columns in the resulting PCA matrix will be named as \verb{\{name\}_1}, \verb{\{name\}_2}, ... .
This ensures compatibility with conversion to a \code{Seurat} object (\code{\link[Seurat:as.Seurat]{Seurat::as.Seurat.SingleCellExperiment()}}).
}
\concept{sc_clustering}
