% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_sample_norm_clustering.R
\name{sce_cc_fn}
\alias{sce_cc_fn}
\title{Assign cell cycle phase to cells.}
\usage{
sce_cc_fn(sce_final_input_qc, cc_genes, data = NULL)
}
\arguments{
\item{sce_final_input_qc}{(\emph{input target}) A \code{SingleCellExperiment} object.}

\item{cc_genes}{(\emph{input target}) A dataframe.}

\item{data}{Passed to \code{\link[Seurat:reexports]{Seurat::as.Seurat()}}.}
}
\value{
A modified \code{sce_final_input_qc} objects with columns appended to \code{colData()}:
\code{phase}, \code{s_score}, \code{g2m_score}, \code{cc_difference}. \emph{Output target}: \code{sce_cc}
}
\description{
This is a wrapper around \code{\link[Seurat:CellCycleScoring]{Seurat::CellCycleScoring()}}.
}
\concept{single_sample_norm_clustering_fn}
