% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_sample_input_qc.R
\name{sce_final_input_qc_fn}
\alias{sce_final_input_qc_fn}
\title{Finalize a \code{SingleCellExperiment} object which will proceed to the \verb{02_norm_clustering} stage.}
\usage{
sce_final_input_qc_fn(sce_selected, gene_annotation, artifact = FALSE)
}
\arguments{
\item{sce_selected}{(\emph{input target}) A \code{SingleCellExperiment} object.}

\item{gene_annotation}{(\emph{input target}) A dataframe.}
}
\value{
A \code{SingleCellExperiment} object. \emph{Output target}: \code{sce_final_input_qc}
}
\description{
Finalize a \code{SingleCellExperiment} object which will proceed to the \verb{02_norm_clustering} stage.
}
\concept{single_sample_input_qc_fn}
