% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hvg.R
\name{sce_get_cc_genes}
\alias{sce_get_cc_genes}
\title{Identify cell cycle-related genes by using variance explained threshold.}
\usage{
sce_get_cc_genes(sce, var_expl_threshold, variable = "phase", ...)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object.}

\item{var_expl_threshold}{A numeric scalar: threshold for variance explained.
Features exceeding this threshold will be marked as CC-related.}

\item{variable}{A character scalar: column to use for calculation of variance explained.}

\item{...}{Passed to \code{\link[scater:getVarianceExplained]{scater::getVarianceExplained()}}.}
}
\value{
A character vector of gene IDs marked as CC-related.
}
\description{
For more details see the HVG selection section in \verb{02_norm_clustering.yaml} config.
}
\concept{sc_hvg}
