% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_integration.R
\name{sce_int_clustering_df_fn}
\alias{sce_int_clustering_df_fn}
\title{Compute a quick graph-based clustering for each integration method.}
\usage{
sce_int_clustering_df_fn(
  sce_int_pca_df,
  snn_k = 10,
  snn_type = "rank",
  snn_clustering_method = c("walktrap", "louvain"),
  BPPARAM = BiocParallel::SerialParam()
)
}
\arguments{
\item{sce_int_pca_df}{(\emph{input target}) A tibble.}

\item{snn_k, snn_type, snn_clustering_method}{See \code{\link[=sce_int_raw_snn_clustering_fn]{sce_int_raw_snn_clustering_fn()}}.}

\item{BPPARAM}{A \link[BiocParallel:BiocParallelParam-class]{BiocParallel::BiocParallelParam} object.}
}
\value{
A modified \code{sce_int_pca_df} with appended \code{SingleCellExperiment} object with computed clustering.
\emph{Output target}: \code{sce_int_clustering_df}

The following items of \code{metadata()} of each integrated \code{SingleCellExperiment} object are added or modified:
\itemize{
\item \code{int_cluster_snn_k}, \code{int_cluster_snn_type}, \code{int_cluster_snn_method}:
taken from parameters of the integration method.
}
}
\description{
For Harmony integration, the \code{"harmony"} reduced dims are used as an input to \code{\link[scran:buildSNNGraph]{scran::buildSNNGraph()}};
otherwise PCA is used.
}
\concept{integration_integration_fn}
