% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_integration.R
\name{sce_int_combine_hvgs}
\alias{sce_int_combine_hvgs}
\title{Combine HVG data from a list of \code{SingleCellExperiment} objects.}
\usage{
sce_int_combine_hvgs(
  sce_list,
  hvg_selection_value = 0,
  hvg_metric_fit_name = "hvg_metric_fit",
  hvg_combination = c("hvg_metric", "intersection", "union", "all"),
  hvg_selection = c("top", "significance", "threshold")
)
}
\arguments{
\item{sce_list}{A named list of \code{SingleCellExperiment} objects.}

\item{hvg_selection_value, hvg_selection}{See \code{\link[=get_top_hvgs]{get_top_hvgs()}}.}

\item{hvg_metric_fit_name}{A character scalar: name of a list item with HVG data in \code{metadata()}.}

\item{hvg_combination}{See \code{\link[=sce_int_import_fn]{sce_int_import_fn()}}.}
}
\value{
A named list of combined HVG data.
}
\description{
This is, in fact, similar to \code{\link[=get_top_hvgs]{get_top_hvgs()}}. The only difference is that HVGs are first combined
from all \code{SingleCellExperiment} objects in \code{sce_list} and consequently, top HVGs are selected.
}
\concept{integration_integration_fn}
