% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_integration.R
\name{sce_int_df_fn}
\alias{sce_int_df_fn}
\title{Perform integration of single-samples.}
\usage{
sce_int_df_fn(
  sce_int_multibatchnorm,
  integration_methods_df,
  BPPARAM = BiocParallel::SerialParam()
)
}
\arguments{
\item{sce_int_multibatchnorm}{(\emph{input target}) A \code{SingleCellExperiment} object
processed by \code{\link[batchelor:multiBatchNorm]{batchelor::multiBatchNorm()}}.}

\item{integration_methods_df}{(\emph{input target}) A tibble: transformed from \code{INTEGRATION_METHODS} parameter in
\verb{01_integration.yaml} config.}

\item{BPPARAM}{A \link[BiocParallel:BiocParallelParam-class]{BiocParallel::BiocParallelParam} object.}
}
\value{
A modified \code{integration_methods_df} tibble with appended integrated \code{SingleCellExperiment} object.
\emph{Output target}: \code{sce_int_df}

The following items of \code{metadata()} of each \code{SingleCellExperiment} object are added or modified:
\itemize{
\item \code{single_samples_metadata_df}: a tibble with information about individual samples:
\itemize{
\item \code{sample_name}, \code{description}, \code{cache_path}, \code{n_features}, \code{hvg_rm_cc_genes}, \code{hvg_cc_genes_var_expl_threshold},
\code{has_filtered_doublets}, \code{max_doublet_score}, \code{Samples}: vectors merged from \code{metadata()} of individual samples.
\item \code{n_cells}: number of cells.
\item \code{hvg_ids}, \code{hvg_rm_cc_genes_ids}: lists of character vectors of HVG / removed HVG IDs.
\item \code{cell_groupings}: list of cell grouping definitions.
\item \code{hvg_metric}: list of HVG metrics.
}
\item \code{hvg_rm_cc_genes}, \code{int_method_name}, \code{int_param}: taken from integration method parameters.
\item \code{hvg_rm_cc_genes_ids}: a character scalar of removed CC-related genes' IDs.
}
}
\description{
Perform integration of single-samples.
}
\concept{integration_integration_fn}
