% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_integration.R
\name{sce_int_dimred_df_fn}
\alias{sce_int_dimred_df_fn}
\title{Calculate t-SNE and UMAP for each integration method result (\code{SingleCellExperiment} object).}
\usage{
sce_int_dimred_df_fn(dimred_params_df, BPPARAM = BiocParallel::SerialParam())
}
\arguments{
\item{dimred_params_df}{(\emph{input target}) A tibble: derived from \code{sce_int_pca_df} and \code{integration_methods_df} targets.
Dimred parameters are defined for each integration method in \code{INTEGRATION_METHODS} parameter
in \verb{01_integration.yaml} config.}

\item{BPPARAM}{A \link[BiocParallel:BiocParallelParam-class]{BiocParallel::BiocParallelParam} object.}
}
\value{
A modified \code{dimred_params_df} tibble with appended \code{SingleCellExperiment} object with computed PCA.
\emph{Output target}: \code{sce_int_dimred_df}
}
\description{
Calculate t-SNE and UMAP for each integration method result (\code{SingleCellExperiment} object).
}
\concept{integration_integration_fn}
