% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_integration.R
\name{sce_int_dimred_plots_df_fn}
\alias{sce_int_dimred_plots_df_fn}
\title{Make dimred plot for each combination of integration method, dimred method, and coloring by
batch (single-sample) and cell cycle phase.}
\usage{
sce_int_dimred_plots_df_fn(dimred_plots_params_df)
}
\arguments{
\item{dimred_plots_params_df}{(\emph{input target}) A tibble.}
}
\value{
A modified \code{dimred_plots_params_df} tibble with appended \code{ggplot2} object.
\emph{Output target}: \code{sce_int_dimred_plots_df}
}
\description{
Make dimred plot for each combination of integration method, dimred method, and coloring by
batch (single-sample) and cell cycle phase.
}
\concept{integration_integration_fn}
