% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_integration.R
\name{sce_int_import_fn}
\alias{sce_int_import_fn}
\title{Import \code{sce_final_norm_clustering} targets from \code{drake} caches.}
\usage{
sce_int_import_fn(
  integration_sources,
  hvg_combination = c("hvg_metric", "intersection", "union", "all")
)
}
\arguments{
\item{integration_sources}{A list of sources to load \code{sce_final_norm_clustering} targets from.
See \verb{01_integration.yaml} config for more details.}

\item{hvg_combination}{A character scalar: method to use for combining HVGs:
\itemize{
\item \code{"hvg_metric"}: combine gene variance or CV2,
e.g. all samples had \code{HVG_METRIC} set to \code{"gene_var"}(\code{\link[scran:combineVar]{scran::combineVar()}}) or
\code{"gene_cv2"} (\code{\link[scran:combineVar]{scran::combineCV2()}}).
\item \code{"intersection"}: make intersection of HVGs.
\item \code{"union"}: make union of HVGs.
\item \code{"all"}: take all HVGs from all samples (without duplicates).
}}
}
\value{
A list of \code{SingleCellExperiment} objects. \emph{Output target}: \code{sce_int_import}

The following items of \code{metadata()} of each imported \code{SingleCellExperiment} object are added or modified:
\itemize{
\item \code{single_sample_cache_path}, \code{single_sample_name}, \code{single_sample_description}, \code{hvg_rm_cc_genes},
\code{hvg_cc_genes_var_expl_threshold}: taken from \code{INTEGRATION_SOURCES} config parameter.
\item \code{hvg_combination}: the value of the \code{hvg_combination} function argument.
}
}
\description{
\code{SingleCellExperiment} objects are also checked for common HVG metric,
and HVG combination method is validated.
}
\concept{integration_integration_fn}
