% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_integration.R
\name{sce_int_processed_fn}
\alias{sce_int_processed_fn}
\title{Subset each object in a list of \code{SingleCellExperiment} objects to common data and their corresponding
metadata (e.g. HVGs).}
\usage{
sce_int_processed_fn(sce_int_raw)
}
\arguments{
\item{sce_int_raw}{(\emph{input target}) A named list of \code{SingleCellExperiment} objects.}
}
\value{
A list of \code{SingleCellExperiment} objects. \emph{Output target}: \code{sce_int_processed}

The following items of \code{metadata()} of each \code{SingleCellExperiment} object are added or modified:
\itemize{
\item \code{hvg_ids}, \code{hvg_metric_fit}: subsetted to common features.
\item \code{n_features_orig}: an integer scalar: number of features (rows) before subsetting to common space.
}
}
\description{
This will subset all \code{SingleCellExperiment} objects to common:
\itemize{
\item Features and HVGs (whose data are saved in metadata).
\item Column data (\code{colData()}). Columns starting with \code{"cluster_"} will be renamed to \code{"cluster_<batch>"},
e.g. \code{"cluster_sc3_3_pbmc1k"}.
}
}
\concept{integration_integration_fn}
