% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_integration.R
\name{sce_int_raw_snn_clustering_fn}
\alias{sce_int_raw_snn_clustering_fn}
\title{Perform a fast shared nearest neighbor clustering of each sample.}
\usage{
sce_int_raw_snn_clustering_fn(
  sce_int_raw,
  snn_k = 10L,
  snn_type = "rank",
  snn_clustering_method = c("walktrap", "louvain"),
  BPPARAM = BiocParallel::SerialParam()
)
}
\arguments{
\item{sce_int_raw}{(\emph{input target}) A named list of \code{SingleCellExperiment} objects.}

\item{snn_k}{An integer scalar: number of nearest neighbors. Passed to \code{\link[scran:buildSNNGraph]{scran::buildSNNGraph()}}.}

\item{snn_type}{A character scalar: type of weighting scheme to use for shared neighbors.
Passed to \code{\link[scran:buildSNNGraph]{scran::buildSNNGraph()}}.}

\item{snn_clustering_method}{A character scalar: type of graph clustering method:
\itemize{
\item \code{"walktrap"}: \code{\link[igraph:cluster_walktrap]{igraph::cluster_walktrap()}}
\item \code{"louvain"}: \code{\link[igraph:cluster_louvain]{igraph::cluster_louvain()}}
}}

\item{BPPARAM}{A \link[BiocParallel:BiocParallelParam-class]{BiocParallel::BiocParallelParam} object.}
}
\value{
A list of \code{SingleCellExperiment} objects. \emph{Output target}: \code{sce_int_raw_snn_clustering}
}
\description{
Clustering is used for integration diagnostics.
}
\concept{integration_integration_fn}
