% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_sample_norm_clustering.R
\name{sce_norm_hvg_fn}
\alias{sce_norm_hvg_fn}
\title{Find highly variable genes (HVGs).}
\usage{
sce_norm_hvg_fn(
  sce_norm,
  hvg_selection_value,
  hvg_metric = c("gene_var", "gene_cv2", "sctransform"),
  hvg_selection = c("top", "significance", "threshold"),
  hvg_rm_cc_genes = FALSE,
  hvg_cc_genes_var_expl_threshold = 5,
  svg = FALSE,
  BSPARAM = BiocSingular::IrlbaParam(),
  BPPARAM = BiocParallel::SerialParam()
)
}
\arguments{
\item{sce_norm}{(\emph{input target}) A \code{SingleCellExperiment} object.}

\item{hvg_selection_value, hvg_metric, hvg_selection}{Passed to \code{\link[=get_top_hvgs]{get_top_hvgs()}} - see its help page.}

\item{hvg_rm_cc_genes}{A logical scalar: if \code{TRUE}, remove cell cycle-related genes exceeding the
\code{hvg_cc_genes_var_expl_threshold} prior to HVG selection.}

\item{hvg_cc_genes_var_expl_threshold}{A numeric scalar: threshold for variance explained.
Genes exceeding this threshold will be marked as CC-related.}

\item{BSPARAM}{A \link[BiocSingular:BiocSingularParam]{BiocSingular::BiocSingularParam} object.}

\item{BPPARAM}{A \link[BiocParallel:BiocParallelParam-class]{BiocParallel::BiocParallelParam} object.}

\item{spatial}{A logical scalar: if \code{TRUE}, add spatially variable genes extension}
}
\value{
A modified \code{sce_norm} object with added HVG data in \code{metadata()}.
\emph{Output target}: \code{sce_norm_hvg}

The following items of \code{metadata(sce_norm)} are added or modified:
\itemize{
\item \code{hvg_metric}, \code{hvg_selection}, \code{hvg_selection_value}, \code{hvg_rm_cc_genes}, \code{hvg_cc_genes_var_expl_threshold}:
values passed to the function.
\item \code{hvg_metric_fit}: either \code{DataFrame} returned from \code{\link[scran:modelGeneVar]{scran::modelGeneVar()}} or \code{\link[scran:modelGeneCV2]{scran::modelGeneCV2()}},
or \code{feature.attributes} dataframe returned from \code{\link[Seurat:SCTransform]{Seurat::SCTransform()}}.
\item Values added by \code{\link[=sce_remove_cc_genes]{sce_remove_cc_genes()}}.
\item \code{hvg_ids}: a character vector of HVG ENSEMBL IDs.
}
}
\description{
Prior to HVG selection, cell cycle-related gene can be removed.
For more details see the HVG selection section in \verb{02_norm_clustering.yaml} config.
}
\concept{single_sample_norm_clustering_fn}
