% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_sample_input_qc.R
\name{sce_raw_fn}
\alias{sce_raw_fn}
\title{Import scRNA-seq data.}
\usage{
sce_raw_fn(input_data, input_data_subset = NULL)
}
\arguments{
\item{input_data}{A named list of named lists containing character scalars:
\itemize{
\item \code{type}:
\itemize{
\item \code{"cellranger"}: a raw feature-barcode matrix from 10x Genomics \code{cellranger}
\item \code{"table"}: a delimited text file (table)
\item \code{"sce"}: a \code{SingleCellExperiment} object (Rds file)
\item \code{"sce_drake_cache"}: a \code{SingleCellExperiment} object loaded from a \code{drake} cache
}
\item \code{path}: a path to input file (or directory in case of \code{type = "cellranger"})
\item \code{delimiter}: a field delimiter when \code{type = "table"}
\item \code{target_name}: a name of \code{SingleCellExperiment} target when \code{type = "sce_drake_cache"}
}}

\item{input_data_subset}{\code{NULL} or a named list by which subsetting of the imported data will be performed.
The list must contain the following items:
\itemize{
\item \code{subset_by} (character scalar): name of column in \code{colData()} to use for subsetting
\item \code{values} (vector): values to subset to
\item \code{negate} (logical scalar): if \code{TRUE}, negate the selection
}}
}
\value{
A \code{SingleCellExperiment} object. \emph{Output target}: \code{sce_raw}
}
\description{
For more details see the \code{INPUT_DATA} parameter in \code{vignette("stage_input_qc")}.
}
\concept{single_sample_input_qc_fn}
