% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hvg.R
\name{sce_remove_cc_genes}
\alias{sce_remove_cc_genes}
\title{Remove cell cycle-related genes from HVGs.}
\usage{
sce_remove_cc_genes(
  sce,
  var_expl_threshold,
  variable = "phase",
  hvg_metric_fit = NULL,
  ...
)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object.}

\item{var_expl_threshold}{A numeric scalar: threshold for variance explained.
Genes exceeding this threshold will be marked as CC-related.}

\item{variable}{A character scalar: column to use for variance explained computation.}

\item{hvg_metric_fit}{A \code{DataFrame} object.}

\item{...}{Passed to \code{\link[=get_top_hvgs]{get_top_hvgs()}}.}
}
\value{
A \code{SingleCellExperiment} object with removed CC-related genes from HVGs.

\code{is_cc_related} logical column is appended to \code{rowData(sce)}.

The following items of \code{metadata(sce)} are added or modified:
\itemize{
\item \code{hvg_cc_genes_var_expl_threshold}: the value of \code{var_expl_threshold}.
\item \code{hvg_ids}: if CC-related genes are found, their IDs are removed.
\item \code{hvg_rm_cc_genes_ids}: IDs of CC-related genes. If these are not found, an empty vector.
}

Note that CC-related genes are not removed from \code{metadata(sce)$hvg_metric_fit}.
}
\description{
Remove cell cycle-related genes from HVGs.
}
\concept{sc_hvg}
