% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_sample_norm_clustering.R
\name{sce_rm_doublets_fn}
\alias{sce_rm_doublets_fn}
\title{Remove cell doublets from a \code{SingleCellExperiment} object.}
\usage{
sce_rm_doublets_fn(sce_norm_hvg, doublet_density, max_doublet_score)
}
\arguments{
\item{sce_norm_hvg}{(\emph{input target}) A \code{SingleCellExperiment} object.}

\item{doublet_density}{(\emph{input target}) A numeric vector.}

\item{max_doublet_score}{A numeric scalar: cells with doublet score (\code{log10(doublet_density + 1)}) exceeding
this value will be marked as doublets and removed from \code{sce_norm_hvg}.
If \code{NULL}, cells won't be filtered out.}
}
\value{
A modified \code{sce_norm_hvg} object with removed cell doublets.

The following items of \code{metadata(sce_norm_hvg)} are added or modified:
\itemize{
\item \code{has_filtered_doublets}: \code{TRUE} if \code{max_doublet_score} is not \code{NULL}.
\item \code{max_doublet_score}: the value of \code{max_doublet_score} function argument.
}
}
\description{
Remove cell doublets from a \code{SingleCellExperiment} object.
}
\concept{single_sample_norm_clustering_fn}
