% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_sample_input_qc.R
\name{sce_selected_fn}
\alias{sce_selected_fn}
\title{Select a \code{SingleCellExperiment} object which will proceed to the \verb{02_norm_clustering} stage.}
\usage{
sce_selected_fn(
  sce_qc_filter_genes,
  sce_custom_filter_genes,
  save_dataset_sensitive_filtering
)
}
\arguments{
\item{sce_qc_filter_genes}{(\emph{input target}) A \code{SingleCellExperiment} object.}

\item{sce_custom_filter_genes}{(\emph{input target}) A \code{SingleCellExperiment} object.}

\item{save_dataset_sensitive_filtering}{A logical scalar: if \code{TRUE}, return \code{sce_qc_filter_genes} target,
otherwise \code{sce_custom_filter_genes} target (both are \code{SingleCellExperiment} objects).}
}
\value{
See the \code{save_dataset_sensitive_filtering} argument. \emph{Output target}: \code{sce_selected}
}
\description{
Select a \code{SingleCellExperiment} object which will proceed to the \verb{02_norm_clustering} stage.
}
\concept{single_sample_input_qc_fn}
