% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_sample_input_qc.R
\name{sce_valid_cells_fn}
\alias{sce_valid_cells_fn}
\title{Subset cells in a \code{SingleCellExperiment} object to non-empty ones and add corresponding statistics.}
\usage{
sce_valid_cells_fn(sce_raw, empty_droplets, empty_droplets_fdr_threshold)
}
\arguments{
\item{sce_raw}{(\emph{input target}) A \code{SingleCellExperiment} object.}

\item{empty_droplets}{(\emph{input target}) A \code{DataFrame} object returned from \code{\link[DropletUtils:emptyDrops]{DropletUtils::emptyDrops()}}.}

\item{empty_droplets_fdr_threshold}{A numeric scalar: threshold value for FDR of a cell being empty.}
}
\value{
A subsetted \code{sce_raw} object. \emph{Output target}: \code{sce_valid_cells}
}
\description{
Subset cells in a \code{SingleCellExperiment} object to non-empty ones and add corresponding statistics.
}
\concept{single_sample_input_qc_fn}
