% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{selected_markers_dimplot}
\alias{selected_markers_dimplot}
\title{Make a grid of feature plots for selected markers.}
\usage{
selected_markers_dimplot(
  sce,
  dimred,
  selected_markers_df,
  assay = "logcounts",
  low_color = "lightgrey",
  high_color = "red",
  combine = TRUE
)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object.}

\item{dimred}{A character scalar: name of dimred to plot.}

\item{selected_markers_df}{A dataframe-like object with two character columns:
\itemize{
\item \code{group}: marker group.
\item \code{markers}: gene symbols of markers separated by \code{":"}.
Such dataframe is created from a CSV file defined in the \code{SELECTED_MARKERS_FILE} parameter in e.g.
\verb{02_norm_clustering.yaml} config.
}

\emph{Input target}: \code{selected_markers_df} or \code{selected_markers_int_df}}

\item{assay}{A character scalar: name of assay in \code{sce} to be used.}

\item{low_color, high_color}{Passed as \code{list(low = low_color, high = high_color)} to \code{\link[=plotReducedDim_mod]{plotReducedDim_mod()}}.}

\item{combine}{A logical scalar: if \code{TRUE}, combine plots to a single \code{patchwork} object, otherwise return
list of \code{ggplot2} objects.}
}
\value{
See the \code{combine} argument.
}
\description{
Make a grid of feature plots for selected markers.
}
\concept{sce_visualization}
