% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_integration.R
\name{selected_markers_int_df_fn}
\alias{selected_markers_int_df_fn}
\title{Prepare parameters for expression plots of selected markers.}
\usage{
selected_markers_int_df_fn(selected_markers_int_file, sce_int_dimred_df)
}
\arguments{
\item{selected_markers_int_file}{A character scalar: path to CSV file with marker definitions.
Defined in \code{SELECTED_MARKERS_INT_FILE} parameter in \verb{01_integration.yaml} config.}

\item{sce_int_dimred_df}{(\emph{input target}) A tibble.}
}
\value{
A modified \code{sce_int_dimred_df} tibble. \emph{Output target}: \code{selected_markers_int_df}
}
\description{
Prepare parameters for expression plots of selected markers.
}
\concept{integration_integration_fn}
