% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_sample_norm_clustering.R
\name{selected_markers_plots_fn}
\alias{selected_markers_plots_fn}
\title{Make a grid of feature plots for selected genes.}
\usage{
selected_markers_plots_fn(sce_final_norm_clustering, selected_markers_df)
}
\arguments{
\item{sce_final_norm_clustering}{(\emph{input target}) A \code{SingleCellExperiment} object.}

\item{selected_markers_df}{(\emph{input target}) A tibble.}
}
\value{
A tibble. \emph{Output target}: \code{selected_markers_plots}
}
\description{
Make a grid of feature plots for selected genes.
}
\concept{single_sample_norm_clustering_fn}
