% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_integration.R
\name{selected_markers_plots_int_df_fn}
\alias{selected_markers_plots_int_df_fn}
\title{Make expression plots of selected markers.}
\usage{
selected_markers_plots_int_df_fn(selected_markers_int_df, sce_int_dimred_df)
}
\arguments{
\item{selected_markers_int_df}{(\emph{input target}) A tibble.}

\item{sce_int_dimred_df}{(\emph{input target}) A tibble.}
}
\value{
A modified \code{selected_markers_int_df} tibble. \emph{Output target}: \code{selected_markers_plots_int_df}
}
\description{
Make expression plots of selected markers.
}
\concept{integration_integration_fn}
