% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_drake_functions.R
\name{selected_spat_fn}
\alias{selected_spat_fn}
\title{select significant ligand-receptor pairings}
\usage{
selected_spat_fn(spatial_all_scores)
}
\arguments{
\item{spatial_all_scores}{Precalculated cell communication scores from LR_data_fn}
}
\description{
Select significant L-R pairing based on communication scores from expression of interacting cells
}
\concept{single_sample_norm_spatial}
