% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_rstudio_drake_cache}
\alias{set_rstudio_drake_cache}
\title{Set \code{rstudio_drake_cache} option.}
\usage{
set_rstudio_drake_cache(dir, verbose = TRUE)
}
\arguments{
\item{dir}{A character scalar: path to directory with \code{drake} cache.}

\item{verbose}{A logical scalar: if \code{TRUE}, be verbose.
The default value is obtained from \code{getOption("scdrake_verbose")}.}
}
\value{
Invisibly \code{TRUE} if \code{dir} exists, \code{FALSE} otherwise.
}
\description{
Value of this option is used internally by \code{drake} to "loadd target under cursor"
(it can be set in \verb{Tools -> Modify Keyboard Shortcuts}).
This shortcut will call \verb{loadd(<name of target under cursor>, cache = getOptions("rstudio_drake_cache"))}.
}
\concept{misc_utils}
