% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce.R
\name{seu_add_metadata}
\alias{seu_add_metadata}
\title{Append data to metadata o a \code{Seurat} object's assay.}
\usage{
seu_add_metadata(seu, assay = "RNA", ...)
}
\arguments{
\item{seu}{A \code{Seurat} object.}

\item{assay}{A character scalar: name of assay.}

\item{...}{Objects to be added}
}
\value{
A modified \code{seu} object with data appended to its \code{assay} metadata.
}
\description{
Append data to metadata o a \code{Seurat} object's assay.
}
\concept{sc_sce}
