% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visualization.R
\name{spatGenePlot2Dsce}
\alias{spatGenePlot2Dsce}
\title{A function for visualization selected genes in pseudotissue visualization}
\usage{
spatGenePlot2Dsce(
  sce,
  sdimx = "pxl_col_in_fullres",
  sdimy = "pxl_row_in_fullres",
  expression_values = c("counts", "logcounts"),
  genes,
  cell_color_gradient = c("blue", "white", "red"),
  gradient_midpoint = NULL,
  gradient_limits = NULL,
  edge_alpha = NULL,
  scale_alpha_with_expression = FALSE,
  point_shape = c("border", "no_border"),
  point_size = 1,
  point_alpha = 1,
  point_border_col = "black",
  point_border_stroke = 0.1,
  show_legend = T,
  legend_text = 8,
  background_color = "white",
  axis_text = 8,
  axis_title = 8,
  cow_n_col = 2,
  cow_rel_h = 1,
  cow_rel_w = 1,
  cow_align = "h"
)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object.}
}
\value{
A ggplot2 object.
}
\description{
Adapted function from Giotto package \link{Dries et al, 2021}, rewrite for use in scdrake package for a SingleCellExperiment object
}
\concept{spatial_visualization}
