% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{spot_deconvolution_fn}
\alias{spot_deconvolution_fn}
\title{Deconvolution of spatial dataset.}
\usage{
spot_deconvolution_fn(spe, sce_ref, sce_ref_label, deconv_method = "CARD")
}
\arguments{
\item{spe}{(\emph{input target}) A normalized \code{SpatialExperiment} object.}

\item{sce_ref}{A \code{SingeCellExperiment} object obtain from config file.}
}
\value{
A modified \code{spe} object.
}
\description{
Deconvulate spots in spatial dataset using CARD or RCTD algorithm.
}
\concept{deconvolution}
