% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_text.R
\name{text_utils_glue}
\alias{text_utils_glue}
\alias{glue0}
\alias{catg0}
\alias{gluec}
\alias{glue0c}
\alias{catn}
\title{Various wrappers around \code{\link[glue:glue]{glue::glue()}} and \code{\link[=cat]{cat()}}.}
\usage{
glue0(..., .envir = parent.frame())

catg0(..., .envir = parent.frame())

gluec(..., .envir = parent.frame())

glue0c(..., .envir = parent.frame())

catn(...)
}
\arguments{
\item{..., .envir}{Passed to \code{\link[glue:glue]{glue::glue()}}.}
}
\description{
Call \code{glue()} without trimming.

Call \code{glue()} without trimming and \code{cat()} the result.

Call \code{glue()} and coerce the result to character.

Call \code{glue()} without trimming and coerce the result to character.

Join a character vector by newlines and cat the result.
}
\concept{misc_text}
