% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_drake_functions.R
\name{top_LR_cell_ints_fn}
\alias{top_LR_cell_ints_fn}
\title{select top 33 significant combined L-R paring}
\usage{
top_LR_cell_ints_fn(selected_spat)
}
\arguments{
\item{selected_spat}{precalculated significant L-R pairing from selected_spat_fn}
}
\description{
select top 33 significant L-R pairing based on combined L-R expression score across cells
}
\concept{single_sample_norm_spatial}
