% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_drake_functions.R
\name{top_LR_ints_fn_new}
\alias{top_LR_ints_fn_new}
\title{select top 33 significant L-R paring}
\usage{
top_LR_ints_fn_new(selected_spat)
}
\arguments{
\item{selected_spat}{Precalculated significant L-R pairing from selected_spat_fn}
}
\description{
select top 33 significant L-R pairing based on expression
}
\concept{single_sample_norm_spatial}
