% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_update.R
\name{update_config_group}
\alias{update_config_group}
\title{Update a group of configs, i.e. pipeline, single-sample, or integration.}
\arguments{
\item{dir}{A character scalar:
\itemize{
\item For \code{update_pipeline_config()}: a path to directory with \code{pipeline.default.yaml} file.
\item For \code{update_single_sample_configs()}: a path to directory with \verb{00_main.default.yaml},
\verb{01_input_qc.default.yaml}, \verb{02_norm_clustering.default.yaml},
\code{cluster_markers.default.yaml}, and \code{contrasts.default.yaml} files.
\item For \code{update_integration_configs()}: a path to directory with \verb{00_main.default.yaml},
\verb{01_integration.default.yaml}, \verb{02_int_clustering.default.yaml},
\code{cluster_markers.default.yaml}, and \code{contrasts.default.yaml} files.
}}

\item{...}{Passed to \code{update_config()}.}
}
\description{
Update a group of configs, i.e. pipeline, single-sample, or integration.
}
