% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{update_project}
\alias{update_project}
\title{Update \code{scdrake} project files.}
\usage{
update_project(
  dir = ".",
  pipeline_config_dir = getOption("scdrake_pipeline_config_dir"),
  single_sample_config_dir = getOption("scdrake_single_sample_config_dir"),
  integration_config_dir = getOption("scdrake_integration_config_dir"),
  ask = TRUE,
  verbose = getOption("scdrake_verbose")
)
}
\arguments{
\item{dir}{A character scalar: path to \code{scdrake} project directory.
If \code{NULL}, the current working directory will be used (same as the default: \code{"."}).}

\item{pipeline_config_dir}{A character scalar: path to directory with pipeline config file (\code{pipeline.yaml}).}

\item{single_sample_config_dir}{A character scalar: path to directory with \verb{00_main.yaml}, \verb{01_input_qc.yaml},
\verb{02_norm_clustering.yaml}, \code{cluster_markers.yaml}, and \code{contrasts.yaml} files.}

\item{integration_config_dir}{A character scalar: path to directory with \verb{00_main.yaml}, \verb{01_integration.yaml},
\verb{02_int_clustering.yaml}, \code{cluster_markers.yaml}, and \code{contrasts.yaml} files.}

\item{ask}{A logical scalar: if \code{TRUE}, ask before updating the files.}

\item{verbose}{A logical scalar: if \code{TRUE}, be verbose.
The default value is obtained from \code{getOption("scdrake_verbose")}.}
}
\value{
Invisibly \code{NULL}.
}
\description{
This will \strong{overwrite} project files by the package-bundled ones:
\itemize{
\item RMarkdown documents in \verb{Rmd/}
\item Initial scripts for \code{\link[drake:r_make]{drake::r_make()}}: \verb{_drake_single_sample.R} and \verb{_drake_integration.R}
\item Default YAML configs. Their location can be set by \verb{*_config_dir} parameters.
}
}
