% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visualization.R
\name{visualized_spots}
\alias{visualized_spots}
\title{Spot visualization}
\usage{
visualized_spots(
  sce,
  sdimx = "Dims_x",
  sdimy = "Dims_y",
  spat_enr_names = NULL,
  cell_color = NULL,
  color_as_factor = F,
  cell_color_code = NULL,
  cell_color_gradient = c("navy", "lightcyan", "red"),
  gradient_midpoint = NULL,
  gradient_limits = NULL,
  select_cells = NULL,
  point_shape = c("border", "no_border"),
  point_size = 3,
  point_alpha = 1,
  point_border_col = "black",
  point_border_stroke = 0.1,
  show_cluster_center = F,
  show_center_label = F,
  center_point_size = 4,
  center_point_border_col = "black",
  center_point_border_stroke = 0.1,
  label_size = 4,
  label_fontface = "bold",
  show_other_cells = T,
  other_cell_color = "lightgrey",
  other_point_size = 1,
  other_cells_alpha = 0.1,
  coord_fix_ratio = NULL,
  title = NULL,
  show_legend = T,
  legend_text = 8,
  legend_symbol_size = 1,
  background_color = "white",
  axis_text = 8,
  axis_title = 8
)
}
\arguments{
\item{sce}{A \code{SingleCellExpression} object}

\item{sdimx}{A label of x coordinates of spots}

\item{sdimy}{A label of y coordinates of spots}

\item{cell_color_gradient}{A vector of length three with color names}
}
\description{
Main function for visualization spots as pseudotissue plot  and subsequent information
}
\concept{spatial_visualization}
