% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce.R
\name{which_genes_regex}
\alias{which_genes_regex}
\title{Get indices of genes whose annotation matches a regex.}
\usage{
which_genes_regex(sce, regex, colname = "Symbol", ignore_case = TRUE)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object.}

\item{regex}{A character scalar.}

\item{colname}{A character scalar: name of column in \code{rowData(sce)} whose values will be tested by \code{regex}.}

\item{ignore_case}{A logical scalar: if \code{TRUE}, ignore case in \code{regex}.}
}
\value{
An integer vector: indices of \code{rowData(sce)} matching the \code{regex}.
}
\description{
Get indices of genes whose annotation matches a regex.
}
\examples{
\dontrun{
# Indices of mitochondrial genes.
which_genes_regex(sce, regex = "^MT-", colname = "Symbol", ignore_case = TRUE)
}

}
\concept{sc_sce}
