% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{with_dbi}
\alias{with_dbi}
\title{Load a SQL database file and run a function from the \code{AnnotationDbi} package.}
\usage{
with_dbi(annotation_db_file, dbi_fun, ...)
}
\arguments{
\item{annotation_db_file}{A character scalar: path to SQLite file of annotation DB, e.g. \code{org.Hs.eg.db$conn@dbname}.}

\item{dbi_fun}{A function from the \code{AnnotationDbi} package which accepts \code{AnnotationDbi} object as parameter \code{x},
e.g. \code{\link[AnnotationDbi:AnnotationDb-class]{AnnotationDbi::mapIds()}}.}

\item{...}{Parameters passed to \code{dbi_fun}.}
}
\value{
An output from \code{dbi_fun}.
}
\description{
An \code{AnnotationDbi} object cannot be used in parallel, so this is a workaround.
}
\concept{sc_annotation}
