% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yq.R
\name{yq_merge_cmd}
\alias{yq_merge_cmd}
\title{Merge two YAML files using the \code{yq} tool.}
\usage{
yq_merge_cmd(
  f1,
  f2,
  stdout = TRUE,
  check_yq = TRUE,
  check_exit = TRUE,
  yq_binary = getOption("scdrake_yq_binary")
)
}
\arguments{
\item{f1, f2}{A character scalar: path to YAML file. \code{f2} is the file whose parameters will overwrite those in \code{f1}.}

\item{stdout}{A character or logical scalar, or \code{NULL}: passed to \code{\link[base:system2]{base::system2()}}.
\itemize{
\item For character scalar: path to file into which the resulting merged YAML file will be written.
If empty string (\code{""}), the result will be written to R console.
\item For logical scalar \code{TRUE}: return the stdout as a character scalar.
\item For logical scalar \code{FALSE}, or \code{NULL}: discard the stdout.
}}

\item{check_yq}{A logical scalar: if \code{TRUE}, check \code{yq} tool presence with \code{\link[=check_yq]{check_yq()}} prior to merging.}

\item{check_exit}{A logical scalar: if \code{TRUE} and \code{stdout} is \code{FALSE},
check exit code of shell command used for merging.}

\item{yq_binary}{A character scalar: path to \code{yq} tool's binary.}
}
\value{
An output from \code{\link[base:system2]{base::system2()}} - depends on the \code{stdout} argument.
}
\description{
Merge two YAML files using the \code{yq} tool.
}
\concept{yq_tool}
