% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{calc_sc3}
\alias{calc_sc3}
\title{Run SC3 clustering for a specific \code{k}.}
\usage{
calc_sc3(sce_pca, sc3_k, is_integration, BPPARAM = BiocParallel::SerialParam())
}
\arguments{
\item{sce_pca}{A \code{SingleCellExperiment} object with calculated PCA.}

\item{sc3_k}{An integer vector: numbers of clusters for SC3.}

\item{is_integration}{A logical scalar: if \code{TRUE}, clusterings will be named as \verb{cluster_int_*}, otherwise \verb{cluster_*}.}

\item{BPPARAM}{A \link[BiocParallel:BiocParallelParam-class]{BiocParallel::BiocParallelParam} object.}
}
\value{
A \code{SingleCellExperiment} object as returned from \code{\link[SC3:sc3]{SC3::sc3()}}.
}
\description{
See \code{\link[SC3:sc3]{SC3::sc3()}} for more details.
}
\details{
If \code{is_integration} is \code{TRUE}, \code{logcounts(sce_pca)} will be coerced to matrix.
}
\concept{sc_clustering}
