% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_annotation.R
\name{calculate_metadata}
\alias{calculate_metadata}
\title{Calculate metadata for manual cell/spot annotation for heatmap visualisation.}
\usage{
calculate_metadata(sce, enrichment, clustering)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object}

\item{enrichment}{precalculated enrichment score for each cell/spot}
}
\description{
Calculate metadata for manual cell/spot annotation for heatmap visualisation.
}
\concept{manual_annotation}
