% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_pkg_installed}
\alias{check_pkg_installed}
\alias{check_qs_installed}
\alias{check_future_installed}
\alias{check_clustermq_installed}
\alias{check_future.callr_installed}
\alias{check_sc3_version}
\alias{check_scdrake_packages}
\title{Check if a package is installed and display an informative message.}
\usage{
check_pkg_installed(pkg, msg = "", verbose = TRUE)

check_qs_installed(verbose = TRUE)

check_future_installed(verbose = TRUE)

check_clustermq_installed(verbose = TRUE)

check_future.callr_installed(verbose = TRUE)

check_sc3_version(verbose = TRUE)

check_scdrake_packages(verbose = TRUE)
}
\arguments{
\item{pkg}{A character scalar: name of package.}

\item{msg}{A character scalar: additional message to be displayed.}
}
\value{
\code{TRUE} if package is installed, \code{FALSE} otherwise.

\code{check_scdrake_packages()} returns \code{TRUE} if all tested packages are installed and \code{{SC3}} is installed from
<github.com/gorgitko/SC3>, \code{FALSE} otherwise.
}
\description{
Check if a package is installed and display an informative message.
}
\examples{
check_pkg_installed("utils")
check_pkg_installed("zzz")
check_qs_installed()
}
\concept{checks}
