% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{dimred_plots_clustering_fn}
\alias{dimred_plots_clustering_fn}
\title{Make a dimred plot for each clustering and dimred method.}
\usage{
dimred_plots_clustering_fn(
  sce_dimred,
  dimred_names,
  cluster_df,
  spatial = FALSE,
  out_dir = NULL
)
}
\arguments{
\item{sce_dimred}{A \code{SingleCellExperiment} object with computed dimreds specified in \code{dimred_names}.}

\item{dimred_names}{A character vector: dimred names to use for plotting.}

\item{cluster_df}{A tibble.}

\item{spatial}{A logical vector, TRUE for enable pseudotissue visualization for spatial transcriptomics datasets}

\item{out_dir}{A character scalar: output directory in which PDF and PNG files will be saved.}
}
\value{
A tibble. \emph{Output target}: \code{dimred_plots_clustering}
}
\description{
Make a dimred plot for each clustering and dimred method.
}
\concept{sce_visualization}
