% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{generate_dimred_plots_section}
\alias{generate_dimred_plots_section}
\title{Generate a section with dimred plots used in some RMarkdown files.}
\usage{
generate_dimred_plots_section(
  dimred_names,
  clustering_names,
  selected_markers,
  dimred_plots_clustering,
  dimred_plots_other_vars,
  selected_markers_files_rel_start = ".",
  main_header = "Dimensionality reduction",
  main_header_level = 1
)
}
\arguments{
\item{dimred_names}{A character vector: names of dimreds to output.}

\item{clustering_names}{A character vector: names of clusterings to output.}

\item{selected_markers}{A \code{tibble}.}

\item{dimred_plots_clustering}{A \code{tibble}.}

\item{dimred_plots_other_vars}{A \code{tibble}.}

\item{selected_markers_files_rel_start}{A character scalar: relative start of path to selected markers PDF.
See the \emph{Relative links} section in \code{\link[=create_a_link]{create_a_link()}}.}

\item{main_header}{A character scalar: text of the main header.}

\item{main_header_level}{A numeric scalar: level of the main header.}
}
\value{
Invisibly \code{NULL}.
}
\description{
Generate a section with dimred plots used in some RMarkdown files.
}
\concept{misc_html}
