% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visualization.R
\name{getDistinctColors}
\alias{getDistinctColors}
\title{A helper function for asigning colors in pseudotissue visualization}
\usage{
getDistinctColors(n)
}
\arguments{
\item{n}{Number of desired colors.}
}
\value{
A character vector of distinct colors.
}
\description{
Adapted function from Giotto package \link{Dries et al, 2021}, rewrite for use in scdrake package for a SingleCellExperiment object
}
\concept{spatial_visualization}
