% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visualization.R
\name{plot_spat_point_layer_ggplot}
\alias{plot_spat_point_layer_ggplot}
\title{A function for pseudotissue visualization}
\usage{
plot_spat_point_layer_ggplot(
  ggobject,
  sdimx = NULL,
  sdimy = NULL,
  cell_locations_metadata_selected,
  cell_locations_metadata_other,
  cell_color = NULL,
  color_as_factor = T,
  cell_color_code = NULL,
  cell_color_gradient = c("yellow", "white", "red"),
  gradient_midpoint = NULL,
  gradient_limits = NULL,
  select_cells = NULL,
  point_size = 2,
  point_alpha = 1,
  point_border_col = "lightgrey",
  point_border_stroke = 0.1,
  label_size = 4,
  label_fontface = "bold",
  show_other_cells = T,
  other_cell_color = "lightgrey",
  other_point_size = 1,
  show_legend = TRUE
)
}
\arguments{
\item{ggobject}{An inheriated object from visualized_spots.}

\item{cell_locations_metadata_selected, cell_locations_metadata_other, cell_color, ...}{Inheriated, passed to ggplot2 object.}
}
\value{
A \code{ggplot2} object.
}
\description{
Adapted function from Giotto package \link{Dries et al, 2021}, rewrite for use in scdrake package in a SingleCellExperiment object.
}
\concept{spatial_visualization}
