% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visualization.R
\name{plot_spat_visuals}
\alias{plot_spat_visuals}
\title{Control quality spatial plot}
\usage{
plot_spat_visuals(sce)
}
\description{
Function for rendering of quality control plots of selected features
}
\concept{spatial_visualization}
