% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visualization.R
\name{plot_spat_visuals}
\alias{plot_spat_visuals}
\title{A function for visualization selected qc matrices in image}
\usage{
plot_spat_visuals(spe)
}
\arguments{
\item{sce}{A \code{SpatialExperiment} object.}
}
\value{
A list of plots.
}
\description{
Plot spots features with the H&E Visium low resolution image
}
\concept{spatial_visualization}
