% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_pdf}
\alias{save_pdf}
\title{Save a list of plots to multipage PDF.}
\usage{
save_pdf(
  plots,
  output_file,
  width = NULL,
  height = NULL,
  make_thumbnail = FALSE,
  stop_on_error = FALSE
)
}
\arguments{
\item{plots}{A list of plots.}

\item{output_file}{A character scalar: path to output PDF file.
If file's directory doesn't exist, it will be created recursively.}

\item{width, height}{A numeric scalar: default width and height of graphics region in inches. Defaults to 7.}

\item{make_thumbnail}{A logical scalar: if \code{TRUE}, a PNG file will be created from the first plot in \code{plots}.}

\item{stop_on_error}{A logical scalar: if \code{TRUE}, the function will stop when there is an error during the
saving of the plot. Otherwise a dummy PDF/PNG thumbnail file with error description will be created instead.}
}
\value{
A named list with the following items:
\itemize{
\item \code{success}: a logical scalar indicating whether the plot saving succeeded (\code{TRUE}) or not (\code{FALSE}).
\item \code{error}: a character scalar with error message if \code{success} is \code{FALSE}, \code{NULL} otherwise.
\item \code{error_plot}: a \code{ggplot2} object with error plot if \code{success} is \code{FALSE}, \code{NULL} otherwise.
\item \code{output_file}: a character scalar, identical to \code{output_file} parameter.
\item \code{thumbnail_file}: a character scalar, path to thumbnail PNG file if \code{make_thumbnail} is \code{TRUE}, \code{NULL} otherwise.
}

Note that \code{success} of \code{FALSE} and the accompanying error message and plot are only possible when \code{stop_on_error} is \code{FALSE}.
}
\description{
Save a list of plots to multipage PDF.
}
\concept{misc_utils}
