% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdrake-package.R
\docType{package}
\name{scdrake-package}
\alias{scdrake}
\alias{scdrake-package}
\title{scdrake: A pipeline for droplet-based single-cell RNA-seq data secondary analysis implemented in the drake Make-like toolkit for R language}
\description{
This pipeline is implemented in the drake Make-like toolkit (https://github.com/ropensci/drake) and provides the following steps of scRNA-seq secondary analysis: input quality control, filtering of cells/genes, normalization, clustering, cell type annotation, cluster markers, differential expression analysis (constrasts -> cluster vs. cluster), integration of multiple samples. All main parts are accompanied by rich graphical outputs and HTML reports (RMarkdown). Thanks to the drake package, the pipeline is highly reproducible, scalable and efficient, provides easy access to all intermediate results, and can be arbitrarily extended as pipeline definitions are abstracted as R objects.
}
\section{Package options}{
 See \code{\link[=get_scdrake_default_options]{get_scdrake_default_options()}}.
}

\section{Included datasets}{

For pipeline testing two datasets from 10x Genomics can be automatically downloaded:
\itemize{
\item \code{\link[=download_pbmc1k]{download_pbmc1k()}}: \href{https://support.10xgenomics.com/single-cell-gene-expression/datasets/3.0.0/pbmc_1k_v3?}{PBMC 1k}
(v3 chemistry, Cell Ranger 3.0.0)
\item \code{\link[=download_pbmc3k]{download_pbmc3k()}}: \href{https://support.10xgenomics.com/single-cell-gene-expression/datasets/1.1.0/pbmc3k}{PBMC 3k}
(v1 chemistry, Cell Ranger 1.1.0). This dataset is used for testing of integration pipeline.
}

\strong{All credits for these datasets go to 10x Genomics.}
Visit \url{https://www.10xgenomics.com/resources/datasets} for more information.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bioinfocz/scdrake}
  \item \url{https://bioinfocz.github.io/scdrake}
  \item Report bugs at \url{https://github.com/bioinfocz/scdrake/issues}
}

}
\author{
\strong{Maintainer}: Jiri Novotny \email{jiri.novotny@img.cas.cz} (\href{https://orcid.org/0000-0003-1338-638X}{ORCID})

Authors:
\itemize{
  \item Jan Kubovciak \email{jan.kubovciak@img.cas.cz}
}

}
\keyword{internal}
