% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce.R
\name{sce_add_colData}
\alias{sce_add_colData}
\title{Append new columns to \code{colData} of a \code{SingleCellExperiment} object.}
\usage{
sce_add_colData(sce, ...)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object.}

\item{...}{Unnamed arguments: dataframe-like objects.
Named arguments (names will be used for new columns): vectors.}
}
\value{
A modified \code{sce} object with added columns.
}
\description{
Append new columns to \code{colData} of a \code{SingleCellExperiment} object.
}
\examples{
\dontrun{
df <- data.frame(example = rownames(colnames))
sce <- sce_add_colData(sce, df, new_column = sce$Barcode)
}

}
\concept{sc_sce}
