% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce.R
\name{sce_add_spatial_colData}
\alias{sce_add_spatial_colData}
\title{Append new columns to \code{colData} of a \code{SingleCellExperiment} or \code{SpatialExperiment} object.}
\usage{
sce_add_spatial_colData(sce, df, replace = TRUE, spatial)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object.}

\item{df}{A dataframe which will be binded column-wise to \code{colData(sce)}.}

\item{replace}{A logical scalar: if \code{TRUE} and columns to be added already exist, they will be first removed
from \code{colData(sce)}.}

\item{spatial}{A logical scalar: if \code{TRUE} add spatial quality control results}
}
\value{
A modified \code{sce} object with added columns.
}
\description{
Append new columns to \code{colData} of a \code{SingleCellExperiment} or \code{SpatialExperiment} object.
}
\examples{
\dontrun{
df <- data.frame(example = rownames(colnames))
sce <- sce_add_colData(sce, df = df)
}

}
\concept{sc_sce}
