% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{sce_compute_dimreds}
\alias{sce_compute_dimreds}
\title{Compute t-SNE and UMAP dimreds on a \code{SingleCellExperiment} object.}
\usage{
sce_compute_dimreds(
  sce_pca_selected_pcs,
  tsne_perp,
  tsne_max_iter,
  dimred = "pca",
  n_dimred = "selected_pcs",
  BPPARAM = BiocParallel::SerialParam()
)
}
\arguments{
\item{sce_pca_selected_pcs}{A \code{SingleCellExperiment} object with calculated PCA and known number of selected PCs.}

\item{tsne_perp}{A numeric scalar: t-SNE perplexity.}

\item{tsne_max_iter}{A numeric scalar: number of t-SNE iterations.}

\item{dimred}{A character scalar: name of matrix in \code{reducedDim()} used to calculate the dimreds.}

\item{n_dimred}{\itemize{
\item A character scalar \code{"selected_pcs"}: if \code{dimred} is \code{"pca"}, then use selected PCs stored in \code{metadata(sce_pca_selected_pcs)$pca_selected_pcs}
\item \code{NULL}: use all dimensions of \code{dimred}
\item An integer scalar: use first \code{n_dimred} dimensions of \code{dimred}
}}

\item{BPPARAM}{A \link[BiocParallel:BiocParallelParam-class]{BiocParallel::BiocParallelParam} object.}
}
\value{
A \code{SingleCellExperiment} object with calculated t-SNE and UMAP dimreds. Column names of matrices of these
dimreds will be named \verb{tsne_<i>} and \verb{umap_<i>}, respectively.
}
\description{
Compute t-SNE and UMAP dimreds on a \code{SingleCellExperiment} object.
}
\concept{sc_clustering}
