% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_sample_input_qc.R
\name{sce_history_plot_fn}
\alias{sce_history_plot_fn}
\title{Plot history of cell and gene filtering.}
\usage{
sce_history_plot_fn(sce_history, spatial = FALSE)
}
\arguments{
\item{sce_history}{(\emph{input target}) A tibble.}

\item{spatial}{A logical vector: \code{TRUE} for spatial transcriptomics dataset.}
}
\value{
A \code{patchwork} object. \emph{Output target}: \code{sce_history_plot}
}
\description{
Plot history of cell and gene filtering.
}
\concept{single_sample_input_qc_fn}
