% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_sample_input_qc.R
\name{sce_raw_fn}
\alias{sce_raw_fn}
\title{Import scRNA-seq data.}
\usage{
sce_raw_fn(input_data)
}
\arguments{
\item{input_data}{A named list of named lists containing character scalars:
\itemize{
\item \code{type}:
\itemize{
\item \code{"cellranger"}: a raw feature-barcode matrix from 10x Genomics \code{cellranger}.
\item \code{"table"}: a delimited text file (table).
\item \code{"sce"}: a \code{SingleCellExperiment} object (Rds file).
}
\item \code{path}: a path to input file (or directory in case of \code{type = "cellranger"}).
\item \code{delimiter}: a field delimiter when \code{type = "table"}.
}}
}
\value{
A \code{SingleCellExperiment} object. \emph{Output target}: \code{sce_raw}
}
\description{
For more details see the \code{INPUT_DATA} parameter in \code{vignette("stage_input_qc")}.
}
\concept{single_sample_input_qc_fn}
