% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visualization.R
\name{spatGenePlot2Dsce}
\alias{spatGenePlot2Dsce}
\title{Spatial gene plot}
\usage{
spatGenePlot2Dsce(
  sce,
  sdimx = "Dims_x",
  sdimy = "Dims_y",
  expression_values = c("counts", "logcounts"),
  genes,
  cell_color_gradient = c("blue", "white", "red"),
  gradient_midpoint = NULL,
  gradient_limits = NULL,
  edge_alpha = NULL,
  scale_alpha_with_expression = FALSE,
  point_shape = c("border", "no_border"),
  point_size = 1,
  point_alpha = 1,
  point_border_col = "black",
  point_border_stroke = 0.1,
  show_legend = T,
  legend_text = 8,
  background_color = "white",
  axis_text = 8,
  axis_title = 8,
  cow_n_col = 2,
  cow_rel_h = 1,
  cow_rel_w = 1,
  cow_align = "h"
)
}
\arguments{
\item{sce}{A \code{SingleCellExpression} object}

\item{sdimx}{A label of x spot coordinates}

\item{sdimy}{A label of y spot coordinates}

\item{genes}{A vector of genes to visualization}

\item{cell_color_gradient}{A vector of lenght three with color names to use as gradient}

\item{values}{A label of which values to use, counts of logcounts}
}
\description{
Spatial visualization of selected genes within pseudotissue representaiton
}
\concept{spatial_visualization}
