% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visualization.R
\name{visualized_spots}
\alias{visualized_spots}
\title{A basic function for pseudotissue visualization}
\usage{
visualized_spots(
  sce,
  sdimx = "Dims_x",
  sdimy = "Dims_y",
  spat_enr_names = NULL,
  cell_color = NULL,
  color_as_factor = F,
  cell_color_code = NULL,
  cell_color_gradient = c("navy", "lightcyan", "red"),
  gradient_midpoint = NULL,
  gradient_limits = NULL,
  select_cells = NULL,
  point_shape = c("border", "no_border"),
  point_size = 3,
  point_alpha = 1,
  point_border_col = "black",
  point_border_stroke = 0.1,
  label_size = 4,
  label_fontface = "bold",
  show_other_cells = T,
  other_cell_color = "lightgrey",
  other_point_size = 1,
  other_cells_alpha = 0.1,
  coord_fix_ratio = NULL,
  title = NULL,
  show_legend = T,
  legend_text = 8,
  legend_symbol_size = 1,
  background_color = "white",
  axis_text = 8,
  axis_title = 8
)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object.}

\item{cell_color, color_as_factor, cell_color_code, ...}{Passed to ggplot2 object in plot_spat_point_layer_ggplot function}
}
\value{
A \code{ggplot2} object.
}
\description{
Adapted function from Giotto package \link{Dries et al, 2021}, rewrite for use in scdrake package in a SingleCellExperiment object
}
\concept{spatial_visualization}
